Actualizar archivos estaticos
