+++ assets/index-BB-A5ZwX.js
@@ -1,3 +1,3 @@
-import{j as o,r as u,u as P,g as Z,I as J,C as N,h as H,i as Q,l as X,F as Y,m as ee,o as te,p as oe,q as se,t as re}from"./vendor_react-BO86OV0u.js";/* empty css                    */import{E as ne,C as ae,a as ie,D as ce,S as le}from"./vendor_icons-material-UfJA-gQ2.js";import{u as R,a as B}from"./vendor_@uidotdev-BWvncYKX.js";import{D as L,a as ue,S as D,I as me,b as T,T as w,d as k,B as E,e as de,f as pe,g as O,h as b,L as fe,i as he,j as ge,k as xe,M as je}from"./vendor_material-D1sBINAn.js";import{d as v}from"./vendor_dayjs-DVEZ5JDk.js";import{T as ve}from"./vendor_@marsidev-CuodNBGj.js";import{P as p}from"./vendor_prop-types-CDP1pqNX.js";import"./vendor_imask-DcyeGTa6.js";import{e as ye,L as be,A as Ce}from"./vendor_x-date-pickers-DMUM-PmT.js";import{S as De}from"./vendor_swr-Dyr5jNOL.js";import{C as Ae}from"./vendor_joy-BOOoD6_M.js";import"./vendor_@babel-Brk9hZUz.js";import"./vendor_stylis-YPZU7XtI.js";import"./vendor_@remix-run-BdEXcqS4.js";import"./vendor_scheduler-CylivXx-.js";import"./vendor_system-DdpyrVMS.js";import"./vendor_styled-engine-BqETdHfM.js";import"./vendor_clsx-B-dksMZM.js";import"./vendor_private-theming-CHDA6u9N.js";import"./vendor_use-sync-external-store-D1USJxqC.js";import"./vendor_dequal-CnqicUFC.js";import"./vendor_utils-DGLdK2NL.js";(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))n(r);new MutationObserver(r=>{for(const a of r)if(a.type==="childList")for(const i of a.addedNodes)i.tagName==="LINK"&&i.rel==="modulepreload"&&n(i)}).observe(document,{childList:!0,subtree:!0});function s(r){const a={};return r.integrity&&(a.integrity=r.integrity),r.referrerPolicy&&(a.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?a.credentials="include":r.crossOrigin==="anonymous"?a.credentials="omit":a.credentials="same-origin",a}function n(r){if(r.ep)return;r.ep=!0;const a=s(r);fetch(r.href,a)}})();function Ee(){var n;const[e,t]=R("DialogMessage",null),s=()=>t(null);return e&&e.message&&o.jsxs(L,{open:!0,sx:{zIndex:99999},children:[o.jsxs(ue,{sx:{display:"flex",pr:1},children:[o.jsxs(D,{direction:"row",alignItems:"center",sx:{flex:1},children:[e!=null&&e.error?o.jsx(ne,{color:"error",sx:{mr:1}}):o.jsx(ae,{color:"success",sx:{mr:1}}),(e==null?void 0:e.title)??(e!=null&&e.error?"Algo ha salido mal":"El registro ha sido exitoso")]}),o.jsx(me,{"aria-label":"close",onClick:s,sx:r=>({ml:2,color:r.palette.grey[500]}),children:o.jsx(ie,{})})]}),o.jsx(T,{dividers:!0,children:o.jsx(D,{children:o.jsx(w,{align:"center",children:(n=e==null?void 0:e.message)==null?void 0:n.split(`
+import{j as o,r as u,u as P,g as Z,I as J,C as N,h as H,i as Q,l as X,F as Y,m as ee,o as te,p as oe,q as se,t as re}from"./vendor_react-BJ2hNBYs.js";/* empty css                    */import{E as ne,C as ae,a as ie,D as ce,S as le}from"./vendor_icons-material-Bkoz6gEh.js";import{u as R,a as B}from"./vendor_@uidotdev-zQNWj2Hl.js";import{D as L,a as ue,S as D,I as me,b as T,T as w,d as k,B as E,e as de,f as pe,g as O,h as b,L as fe,i as he,j as ge,k as xe,M as je}from"./vendor_material-GI4Anplu.js";import{d as v}from"./vendor_dayjs-DVEZ5JDk.js";import{T as ve}from"./vendor_@marsidev-BV-MhFy_.js";import{P as p}from"./vendor_prop-types-CDP1pqNX.js";import"./vendor_imask-DcyeGTa6.js";import{e as ye,L as be,A as Ce}from"./vendor_x-date-pickers-BLC6f5Ep.js";import{S as De}from"./vendor_swr-eIHTgCvT.js";import{C as Ae}from"./vendor_joy-Czalg0vi.js";import"./vendor_@babel-BRdSvYXJ.js";import"./vendor_stylis-YPZU7XtI.js";import"./vendor_@remix-run-BdEXcqS4.js";import"./vendor_scheduler-CylivXx-.js";import"./vendor_system-bhe-e0RG.js";import"./vendor_styled-engine-BuVfaJQk.js";import"./vendor_clsx-B-dksMZM.js";import"./vendor_private-theming-Cy3868wu.js";import"./vendor_use-sync-external-store-DurkBIMh.js";import"./vendor_dequal-CnqicUFC.js";import"./vendor_utils-Ds1aBnGi.js";(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))n(r);new MutationObserver(r=>{for(const a of r)if(a.type==="childList")for(const i of a.addedNodes)i.tagName==="LINK"&&i.rel==="modulepreload"&&n(i)}).observe(document,{childList:!0,subtree:!0});function s(r){const a={};return r.integrity&&(a.integrity=r.integrity),r.referrerPolicy&&(a.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?a.credentials="include":r.crossOrigin==="anonymous"?a.credentials="omit":a.credentials="same-origin",a}function n(r){if(r.ep)return;r.ep=!0;const a=s(r);fetch(r.href,a)}})();function Ee(){var n;const[e,t]=R("DialogMessage",null),s=()=>t(null);return e&&e.message&&o.jsxs(L,{open:!0,sx:{zIndex:99999},children:[o.jsxs(ue,{sx:{display:"flex",pr:1},children:[o.jsxs(D,{direction:"row",alignItems:"center",sx:{flex:1},children:[e!=null&&e.error?o.jsx(ne,{color:"error",sx:{mr:1}}):o.jsx(ae,{color:"success",sx:{mr:1}}),(e==null?void 0:e.title)??(e!=null&&e.error?"Algo ha salido mal":"El registro ha sido exitoso")]}),o.jsx(me,{"aria-label":"close",onClick:s,sx:r=>({ml:2,color:r.palette.grey[500]}),children:o.jsx(ie,{})})]}),o.jsx(T,{dividers:!0,children:o.jsx(D,{children:o.jsx(w,{align:"center",children:(n=e==null?void 0:e.message)==null?void 0:n.split(`
 `).map((r,a)=>o.jsxs(u.Fragment,{children:[r.split(/(<strong>.*?<\/strong>)/g).map((i,f)=>{if(i.startsWith("<strong>")&&i.endsWith("</strong>")){const h=i.slice(8,-9);return o.jsx("strong",{children:h},f)}return i}),o.jsx("br",{})]},a))})})}),o.jsx(k,{children:o.jsx(E,{autoFocus:!0,onClick:s,color:"primary",children:"Listo"})})]})}function Pe(){const e=de();return pe(e.breakpoints.down("md"))}const Re="/assets/banner-Ceg1iZAU.jpg",Se="/assets/sm_banner-B-E5hjF5.jpg",q="/assets/footer-kaUWP7Jz.jpg",Ie="/assets/sm_footer-BSTcoRLE.jpg",Fe="/assets/banner-saQe3D5L.jpg",Ne="/assets/sm_banner-C6w1tEDo.jpg",_="/assets/footer-DJ1k-47K.jpg",Be="/assets/sm_footer-7Fm8BNkQ.jpg",Le="/assets/sm_botones-BNRbp7t3.jpg",Te="/assets/sm_botones-KDw0tJMz.jpg",we="/assets/banner-DI12jFB0.jpg",ke="/assets/sm_banner-gRTR_qIz.jpg",M="/assets/footer-CiARUKpK.jpg",Oe="/assets/sm_footer-DEsMDHBP.jpg",qe="/assets/footer-CiARUKpK.jpg";function _e(e,t){switch(e){case"20hr":return t?Se:Re;case"20hr-institucional":return t?Ne:Fe;case"diplomado":return t?ke:we}}function Me(e,t){switch(e){case"20hr":return t?Ie:q;case"20hr-institucional":return t?Be:_;case"diplomado":return t?Oe:M}}function ze(e,t){switch(e){case"20hr":return t?Le:q;case"20hr-institucional":return t?Te:_;case"diplomado":return t?qe:M}}function We(){const[e,t]=R("form",{}),s=e.expires?v(e.expires):null,n=s?v().isAfter(s):!1;return s&&n?[{},t]:[e.values,t]}function Ve(){const{control:e}=P(),t=Z({control:e});u.useEffect(()=>{(async()=>{t.frontDocument=null,t.backDocument=null,localStorage.setItem("form",JSON.stringify({values:t,expires:v().add(1,"hour").format()}))})()},[t])}const $e=["firstName","lastName","identityDocument","documentNumber","countryExpedition","frontDocument","backDocument","birthdate","countryBirth","stateBirth","cityBirth","gender","ethnicity","entityName","typeEntity","phoneNumber","whatsappNumber","email","stateLocation","address","neighborhood","zona","connectivity","continuar_curso_120","processingOfPersonalData"];function Ke(e,t){var s;for(const n of $e)if(e.includes(n)){(s=t.current[n])==null||s.scrollIntoView({behavior:"smooth",block:"center"});break}}const Ue=(e,{onBlur:t,onChange:s})=>{var n;e.target.value=((n=e.target.value)==null?void 0:n.replaceAll(" ",""))??"",t==null||t(e),s==null||s(e)},F=(e,{onBlur:t,onChange:s})=>{var n;e.target.value=((n=e.target.value)==null?void 0:n.toUpperCase())??"",t==null||t(e),s==null||s(e)},Ge=!0,z=!1,Ze={FORM:"https://panel.consultaprevia.co/api/ubicacion",API:"https://panel.consultaprevia.co/api/v1",SERVER:"https://panel.consultaprevia.co"},Je={firstName:"Nombres",lastName:"Apellidos",identityDocument:"Tipo de documento",documentNumber:"Número de documento",countryExpedition:"País de expedición",frontDocument:"Frente del documento",backDocument:"Reverso del documento",birthdate:"Fecha de nacimiento",cityBirth:"Ciudad de nacimiento",gender:"Género",gender_other:"Otro (especifique)",ethnicity:"Etnia",entityName:"Nombre de la entidad u organización que representa",typeEntity:"De los siguientes roles en cual se reconoce?",phoneNumber:"Número de teléfono",whatsappNumber:"Número de WhatsApp",email:"Correo electrónico",cityLocation:"Ciudad de residencia",address:"Dirección de residencia",neighborhood:"Barrio de residencia",stratum:"Estrato",connectivity:"Conectividad",countryBirth:"País de nacimiento",stateBirth:"Departamento de nacimiento",stateLocation:"Departamento de residencia",processingOfPersonalData:"Autorización para el tratamiento de datos personales"};function He(){const[e,t]=R("DialogMessage",null);return{message:e,onClose:()=>t(null),onOpen:({title:r,message:a,error:i=!1})=>t({title:r,message:a,error:i})}}function Qe(e,t=750){const s=v().diff(e,"milliseconds");return Math.max(0,t-s)}function Xe(){const[e,t]=u.useState(!1),[s,n]=u.useState();return{loading:e,start:i=>{t(!0),n(i||v())},finish:i=>{setTimeout(()=>{t(!1),i()},Qe(s))}}}const Ye=Ze.API+"/inscripcion",et={registrar:e=>fetch(Ye+"/registrar",{method:"POST",body:e})};function tt(e){const t=new FormData;for(const s in e)Array.isArray(e[s])?t.append(s,JSON.stringify(e[s])):t.append(s,e[s]);return t}const W=u.forwardRef(function(t,s){const{onChange:n,...r}=t;return o.jsx(J,{...r,mask:"#00 000 0000",definitions:{"#":/[1-9]/},inputRef:s,onAccept:a=>n({target:{name:t.name,value:a}}),overwrite:!0})});W.propTypes={name:p.string.isRequired,onChange:p.func.isRequired};function V({slotProps:e}){B();const{control:t}=P(),{controller:s,field:{...n},formRef:r}=e;return o.jsx(N,{control:t,render:({field:a,fieldState:{error:i}})=>o.jsx(O,{...a,ref:f=>{r&&(r.current[a.name]=f),a.ref(f)},variant:"outlined",error:!!(i!=null&&i.type||i!=null&&i.types),helperText:(i==null?void 0:i.message)??" ",fullWidth:!0,...n,slotProps:{input:{startAdornment:o.jsx(H,{countryCode:"CO",svg:!0}),endAdornment:z,inputComponent:W}}}),...s})}V.propTypes={slotProps:p.shape({controller:p.object.isRequired,field:p.object.isRequired,formRef:p.any}).isRequired};function C({slotProps:e}){B();const{control:t}=P(),{controller:s,field:{InputProps:n,onChange:r,onBlur:a,...i},formRef:f}=e;return o.jsx(N,{control:t,render:({field:h,fieldState:{error:c}})=>{const{onChange:x,onBlur:g}=h;return o.jsx(O,{...h,ref:m=>{f&&(f.current[h.name]=m),h.ref(m)},onChange:m=>(r==null?void 0:r(m,{onChange:x}))||x(m),onBlur:m=>(a==null?void 0:a(m,{onBlur:g,onChange:x}))||g(m),variant:"outlined",error:!!(c!=null&&c.type||c!=null&&c.types),helperText:(c==null?void 0:c.message)??" ",fullWidth:!0,...i,slotProps:{input:{endAdornment:z,...n}}})},...s})}C.propTypes={slotProps:p.shape({controller:p.object.isRequired,field:p.object.isRequired,formRef:p.any}).isRequired};const ot=[{Component:C,controller:{name:"firstName",defaultValue:"",rules:{required:{value:!0,message:"Este campo no puede estar vacio"},pattern:{value:/^[a-zA-ZáéíóúÁÉÍÓÚñÑ\s]+$/,message:"Por favor verifica el nombre"},maxLength:{value:150,message:"El nombre no puede tener más de 150 caracteres"}}},field:{label:"Nombres",required:!0,onChange:F,disabled:!0}},{Component:C,controller:{name:"lastName",defaultValue:"",rules:{required:{value:!0,message:"Este campo no puede estar vacio"},pattern:{value:/^[a-zA-ZáéíóúÁÉÍÓÚñÑ\s]+$/,message:"Por favor verifica el apellido"},maxLength:{value:150,message:"El apellido no puede tener más de 150 caracteres"}}},field:{label:"Apellidos",required:!0,onChange:F,disabled:!0}},{Component:V,controller:{name:"phoneNumber",defaultValue:"",rules:{required:{value:!0,message:"Este campo no puede estar vacio"}}},field:{label:"Número de teléfono - WhatsApp",required:!0}},{Component:C,controller:{name:"email",defaultValue:"",rules:{required:{value:!0,message:"Este campo no puede estar vacio"},pattern:{value:/^[a-zA-Z0-9._%+-]+@(?:[a-zA-Z0-9-]+\.)+[a-zA-Z]{2,}$/,message:"Por favor verifica el correo"},maxLength:{value:300,message:"El correo no puede tener más de 300 caracteres"}}},field:{label:"Correo electrónico",required:!0,onBlur:Ue}}];function st(){return Ve(),null}function rt(){const{curso:e="diplomado"}=Q(),t=u.useRef(null),s=u.useRef({}),n=Pe(),[r,a]=u.useState(Ge),i=_e(e,n),f=Me(e,n),[h,c]=We(),x=X({mode:"onBlur",defaultValues:h}),{onOpen:g}=He(),{loading:m,start:$,finish:S}=Xe(),I=()=>{c({}),setTimeout(()=>location.reload(),750)},K=l=>{$(v());const j=tt({...l,alreadyRegistered:!0});et.registrar(j).then(d=>S(()=>{d.json().then(y=>{d.ok?(g({message:y.message}),I()):(console.log(y.message),g({message:y.message??`Algo ha fallado al registrarse (${d.status}_${d.statusText.toUpperCase()})`,error:!0}))}).catch(()=>{g({message:"No se ha obtenido respuesta del servidor",error:!0}),A()})})).catch(()=>S(()=>g({message:"Ha ocurrido un error en el servidor",error:!0})))},U=l=>{a(!1),x.setValue("turnstile_token",l)},A=()=>{var l;a(!0),(l=t.current)==null||l.reset()},G=l=>{const j=Object.keys(l);let d=j.map(y=>Je[y]).join(", ");j.length>=3&&(d=`${d}`),g({message:`Por favor verifica los campos: 
 `+d,error:!0,title:"El formulario contiene errores"}),Ke(j,s),A()};return o.jsxs(L,{fullScreen:!0,open:!0,children:[o.jsxs(T,{sx:{p:"0px !important"},children:[o.jsx(b,{component:"img",src:i,alt:"Banner",sx:{width:"100%",pb:1.25}}),o.jsx(b,{sx:{px:2},children:o.jsxs(D,{flexDirection:"column",alignItems:"center",children:[o.jsx(w,{variant:"h6",align:"center",sx:{mb:1},children:"Agradecemos su interes en el proceso de formación en Consulta Previa. Las inscripciones al Diplomado actualmente están cerradas. Si desea, le invitamos a dejar sus datos para contactarle en futuras convocatorias de formación en Consulta Previa. Nombre, apellidos, correo electrónico y número de telefono celular."}),o.jsxs(Y,{...x,children:[o.jsx(b,{sx:{minWidth:{md:"400px",xs:"100%"},maxWidth:"100%"},children:ot.map((l,j)=>{const{Component:d}=l;return o.jsx(d,{slotProps:{...l,formRef:s}},j)})}),o.jsx(st,{})]}),o.jsx(b,{component:ve,onSuccess:U,onExpire:A,options:{theme:"light",refreshExpired:"manual"},ref:t,label:`form_button_${window.location.hostname}`,siteKey:"0x4AAAAAAAiIftnN7i8Mr-yd",sx:{mb:"19.91px !important",mt:"0 !important",display:"flex",justifyContent:"center",alignItems:"center"}})]})}),o.jsx(b,{component:"img",src:f,alt:"Banner",sx:{width:"100%"}})]}),o.jsx(k,{sx:{justifyContent:"space-between",backgroundImage:`url(${ze(e,n)})`,backgroundSize:"cover"},children:m?o.jsx(D,{justifyContent:"center",sx:{width:"100%",height:"36.5px"},children:o.jsx(fe,{sx:{width:"100%"}})}):o.jsxs(u.Fragment,{children:[o.jsx(E,{onClick:I,startIcon:o.jsx(ce,{}),sx:{color:window.location.pathname.includes("20hr")?"white":"black"},children:"Limpiar formulario"}),o.jsx(E,{onClick:x.handleSubmit(K,G),endIcon:o.jsx(le,{}),disabled:r,sx:{color:window.location.pathname.includes("20hr")?"white":"black"},children:"Registrarse"})]})})]})}function nt(){return o.jsxs(u.Fragment,{children:[o.jsx(ee,{children:o.jsx(te,{path:"*",element:o.jsx(rt,{})})}),o.jsx(Ee,{})]})}const at=he({cssVariables:!0},ye,ge),it=oe([{path:"/*",element:o.jsx(nt,{})}]);se(document.getElementById("root")).render(o.jsx(u.StrictMode,{children:o.jsx(xe,{theme:{[je]:at},children:o.jsx(be,{dateAdapter:Ce,adapterLocale:"es",children:o.jsx(Ae,{children:o.jsx(De,{value:{revalidateOnMount:!0,refreshInterval:!1,fetcher:async(...e)=>fetch(...e).then(t=>t.ok?t.json():Promise.reject({status:t.status,statusText:t.statusText}))},children:o.jsx(re,{router:it})})})})})}));
