--- assets/vendor_@marsidev-CuodNBGj.js
@@ -1,1 +1,0 @@
-import{r as o,j as B}from"./vendor_react-BO86OV0u.js";var Z=({as:n="div",...u},i)=>B.jsx(n,{...u,ref:i}),_=o.forwardRef(Z),X="https://challenges.cloudflare.com/turnstile/v0/api.js",I="cf-turnstile-script",ee="cf-turnstile",S="onloadTurnstileCallback",M=n=>!!document.getElementById(n),re=({render:n="explicit",onLoadCallbackName:u=S,scriptOptions:{nonce:i="",defer:e=!0,async:p=!0,id:s="",appendTo:y,onError:T,crossOrigin:k=""}={}})=>{let R=s||I;if(M(R))return;let a=document.createElement("script");a.id=R,a.src=`${X}?onload=${u}&render=${n}`,!document.querySelector(`script[src="${a.src}"]`)&&(a.defer=!!e,a.async=!!p,i&&(a.nonce=i),k&&(a.crossOrigin=k),T&&(a.onerror=T,delete window[u]),(y==="body"?document.body:document.getElementsByTagName("head")[0]).appendChild(a))},h={normal:{width:300,height:65},compact:{width:150,height:140},invisible:{width:0,height:0,overflow:"hidden"},flexible:{minWidth:300,width:"100%",height:65},interactionOnly:{width:"fit-content",height:"auto",display:"flex"}};function te(n){if(n!=="invisible"&&n!=="interactionOnly")return n}function ne(n=I){let[u,i]=o.useState(!1);return o.useEffect(()=>{let e=()=>{M(n)&&i(!0)},p=new MutationObserver(e);return p.observe(document,{childList:!0,subtree:!0}),e(),()=>{p.disconnect()}},[n]),u}var E="unloaded",W,oe=new Promise((n,u)=>{W={resolve:n,reject:u},E==="ready"&&n(void 0)}),ie=(n=S)=>(E==="unloaded"&&(E="loading",window[n]=()=>{W.resolve(),E="ready",delete window[n]}),oe),ae=o.forwardRef((n,u)=>{let{scriptOptions:i,options:e={},siteKey:p,onWidgetLoad:s,onSuccess:y,onExpire:T,onError:k,onBeforeInteractive:R,onAfterInteractive:a,onUnsupported:q,onTimeout:A,onLoadScript:F,id:H,style:U,as:G="div",injectScript:L=!0,...J}=n,c=e.size,C=o.useCallback(()=>typeof c>"u"?{}:e.execution==="execute"?h.invisible:e.appearance==="interaction-only"?h.interactionOnly:h[c],[e.execution,c,e.appearance]),[K,x]=o.useState(C()),l=o.useRef(null),[w,$]=o.useState(!1),t=o.useRef(),v=o.useRef(!1),P=H||ee,N=(i==null?void 0:i.id)||I,O=ne(N),D=(i==null?void 0:i.onLoadCallbackName)||S,Q=e.appearance||"always",g=o.useMemo(()=>({sitekey:p,action:e.action,cData:e.cData,callback:r=>{v.current=!0,y==null||y(r)},"error-callback":k,"expired-callback":T,"before-interactive-callback":R,"after-interactive-callback":a,"unsupported-callback":q,theme:e.theme||"auto",language:e.language||"auto",tabindex:e.tabIndex,"response-field":e.responseField,"response-field-name":e.responseFieldName,size:te(c),retry:e.retry||"auto","retry-interval":e.retryInterval||8e3,"refresh-expired":e.refreshExpired||"auto","refresh-timeout":e.refreshTimeout||"auto",execution:e.execution||"render",appearance:e.appearance||"always","feedback-enabled":e.feedbackEnabled||!0,"timeout-callback":A}),[e.action,e.appearance,e.cData,e.execution,e.language,e.refreshExpired,e.responseField,e.responseFieldName,e.retry,e.retryInterval,e.tabIndex,e.theme,e.feedbackEnabled,e.refreshTimeout,p,c]),m=o.useCallback(()=>typeof window<"u"&&!!window.turnstile,[]);return o.useEffect(function(){L&&!w&&re({onLoadCallbackName:D,scriptOptions:{...i,id:N}})},[L,w,i,N]),o.useEffect(function(){E!=="ready"&&ie(D).then(()=>$(!0)).catch(console.error)},[]),o.useEffect(function(){if(!l.current||!w)return;let r=!1;return(async()=>{if(r||!l.current)return;let d=window.turnstile.render(l.current,g);t.current=d,t.current&&(s==null||s(t.current))})(),()=>{r=!0,t.current&&(window.turnstile.remove(t.current),v.current=!1)}},[P,w,g]),o.useImperativeHandle(u,()=>{let{turnstile:r}=window;return{getResponse(){if(!(r!=null&&r.getResponse)||!t.current||!m()){console.warn("Turnstile has not been loaded");return}return r.getResponse(t.current)},async getResponsePromise(d=3e4,V=100){return new Promise((Y,j)=>{let f,z=async()=>{if(v.current&&window.turnstile&&t.current)try{let b=window.turnstile.getResponse(t.current);return f&&clearTimeout(f),b?Y(b):j(new Error("No response received"))}catch(b){return f&&clearTimeout(f),console.warn("Failed to get response",b),j(new Error("Failed to get response"))}f||(f=setTimeout(()=>{f&&clearTimeout(f),j(new Error("Timeout"))},d)),await new Promise(b=>setTimeout(b,V)),await z()};z()})},reset(){if(!(r!=null&&r.reset)||!t.current||!m()){console.warn("Turnstile has not been loaded");return}e.execution==="execute"&&x(h.invisible);try{v.current=!1,r.reset(t.current)}catch(d){console.warn(`Failed to reset Turnstile widget ${t}`,d)}},remove(){if(!(r!=null&&r.remove)||!t.current||!m()){console.warn("Turnstile has not been loaded");return}x(h.invisible),v.current=!1,r.remove(t.current),t.current=null},render(){if(!(r!=null&&r.render)||!l.current||!m()||t.current){console.warn("Turnstile has not been loaded or container not found");return}let d=r.render(l.current,g);return t.current=d,t.current&&(s==null||s(t.current)),e.execution!=="execute"&&x(c?h[c]:{}),d},execute(){if(e.execution!=="execute"){console.warn('Execution mode is not set to "execute"');return}if(!(r!=null&&r.execute)||!l.current||!t.current||!m()){console.warn("Turnstile has not been loaded or container not found");return}r.execute(l.current,g),x(c?h[c]:{})},isExpired(){return!(r!=null&&r.isExpired)||!t.current||!m()?(console.warn("Turnstile has not been loaded"),!1):r.isExpired(t.current)}}},[t,e.execution,c,g,l,m,w,s]),o.useEffect(()=>{O&&!w&&window.turnstile&&$(!0)},[w,O]),o.useEffect(()=>{x(C())},[e.execution,c,Q]),o.useEffect(()=>{!O||typeof F!="function"||F()},[O]),B.jsx(_,{ref:l,as:G,id:P,style:{...K,...U},...J})});ae.displayName="Turnstile";export{ae as T};
