+++ assets/vendor_@uidotdev-zQNWj2Hl.js
@@ -1,0 +1,1 @@
+import{r as s}from"./vendor_react-BJ2hNBYs.js";function S(t,e){window.dispatchEvent(new StorageEvent("storage",{key:t,newValue:e}))}const c=(t,e)=>{const n=JSON.stringify(e);window.localStorage.setItem(t,n),S(t,n)},l=t=>{window.localStorage.removeItem(t),S(t,null)},u=t=>window.localStorage.getItem(t),i=t=>(window.addEventListener("storage",t),()=>window.removeEventListener("storage",t)),d=()=>{throw Error("useLocalStorage is a client-only hook")};function w(t,e){const n=()=>u(t),r=s.useSyncExternalStore(i,n,d),g=s.useCallback(a=>{try{const o=typeof a=="function"?a(JSON.parse(r)):a;o==null?l(t):c(t,o)}catch(o){console.warn(o)}},[t,r]);return s.useEffect(()=>{u(t)===null&&typeof e<"u"&&c(t,e)},[t,e]),[r?JSON.parse(r):e,g]}function m(){const t=s.useRef(0);return t.current++,t.current}export{m as a,w as u};
