+++ assets/vendor_joy-Czalg0vi.js
@@ -1,4 +1,4 @@
-import{j as R,G as Ct,r as B,T as rt}from"./vendor_react-BO86OV0u.js";import{_ as oe,a as p,t as xt}from"./vendor_@babel-Brk9hZUz.js";import{d as Z,c as nt,f as Oe,g as $t}from"./vendor_utils-DGLdK2NL.js";const vt="$$joy";function wt(e){return e==null||Object.keys(e).length===0}function He(e){const{styles:t,defaultTheme:o={}}=e,r=typeof t=="function"?n=>t(wt(n)?o:n):t;return R.jsx(Ct,{styles:r})}const Bt=["values","unit","step"],kt=e=>{const t=Object.keys(e).map(o=>({key:o,val:e[o]}))||[];return t.sort((o,r)=>o.val-r.val),t.reduce((o,r)=>p({},o,{[r.key]:r.val}),{})};function At(e){const{values:t={xs:0,sm:600,md:900,lg:1200,xl:1536},unit:o="px",step:r=5}=e,n=oe(e,Bt),i=kt(t),s=Object.keys(i);function c(m){return`@media (min-width:${typeof t[m]=="number"?t[m]:m}${o})`}function d(m){return`@media (max-width:${(typeof t[m]=="number"?t[m]:m)-r/100}${o})`}function b(m,y){const S=s.indexOf(y);return`@media (min-width:${typeof t[m]=="number"?t[m]:m}${o}) and (max-width:${(S!==-1&&typeof t[s[S]]=="number"?t[s[S]]:y)-r/100}${o})`}function x(m){return s.indexOf(m)+1<s.length?b(m,s[s.indexOf(m)+1]):c(m)}function w(m){const y=s.indexOf(m);return y===0?c(s[1]):y===s.length-1?d(s[y]):b(m,s[s.indexOf(m)+1]).replace("@media","@media not all and")}return p({keys:s,values:i,up:c,down:d,between:b,only:x,not:w,unit:o},n)}function fe(e,t){return t?Z(e,t,{clone:!1}):e}const Ie={xs:0,sm:600,md:900,lg:1200,xl:1536},Xe={keys:["xs","sm","md","lg","xl"],up:e=>`@media (min-width:${Ie[e]}px)`};function J(e,t,o){const r=e.theme||{};if(Array.isArray(t)){const i=r.breakpoints||Xe;return t.reduce((s,c,d)=>(s[i.up(i.keys[d])]=o(t[d]),s),{})}if(typeof t=="object"){const i=r.breakpoints||Xe;return Object.keys(t).reduce((s,c)=>{if(Object.keys(i.values||Ie).indexOf(c)!==-1){const d=i.up(c);s[d]=o(t[c],c)}else{const d=c;s[d]=t[d]}return s},{})}return o(t)}function Et(e={}){var t;return((t=e.keys)==null?void 0:t.reduce((r,n)=>{const i=e.up(n);return r[i]={},r},{}))||{}}function Ft(e,t){return e.reduce((o,r)=>{const n=o[r];return(!n||Object.keys(n).length===0)&&delete o[r],o},t)}function xe(e,t,o=!0){if(!t||typeof t!="string")return null;if(e&&e.vars&&o){const r=`vars.${t}`.split(".").reduce((n,i)=>n&&n[i]?n[i]:null,e);if(r!=null)return r}return t.split(".").reduce((r,n)=>r&&r[n]!=null?r[n]:null,e)}function Ce(e,t,o,r=o){let n;return typeof e=="function"?n=e(o):Array.isArray(e)?n=e[o]||r:n=xe(e,o)||r,t&&(n=t(n,r,e)),n}function F(e){const{prop:t,cssProperty:o=e.prop,themeKey:r,transform:n}=e,i=s=>{if(s[t]==null)return null;const c=s[t],d=s.theme,b=xe(d,r)||{};return J(s,c,w=>{let m=Ce(b,n,w);return w===m&&typeof w=="string"&&(m=Ce(b,n,`${t}${w==="default"?"":nt(w)}`,w)),o===!1?m:{[o]:m}})};return i.propTypes={},i.filterProps=[t],i}function Dt(e){const t={};return o=>(t[o]===void 0&&(t[o]=e(o)),t[o])}const Ht={m:"margin",p:"padding"},Tt={t:"Top",r:"Right",b:"Bottom",l:"Left",x:["Left","Right"],y:["Top","Bottom"]},qe={marginX:"mx",marginY:"my",paddingX:"px",paddingY:"py"},jt=Dt(e=>{if(e.length>2)if(qe[e])e=qe[e];else return[e];const[t,o]=e.split(""),r=Ht[t],n=Tt[o]||"";return Array.isArray(n)?n.map(i=>r+i):[r+n]}),Pe=["m","mt","mr","mb","ml","mx","my","margin","marginTop","marginRight","marginBottom","marginLeft","marginX","marginY","marginInline","marginInlineStart","marginInlineEnd","marginBlock","marginBlockStart","marginBlockEnd"],Re=["p","pt","pr","pb","pl","px","py","padding","paddingTop","paddingRight","paddingBottom","paddingLeft","paddingX","paddingY","paddingInline","paddingInlineStart","paddingInlineEnd","paddingBlock","paddingBlockStart","paddingBlockEnd"];[...Pe,...Re];function he(e,t,o,r){var n;const i=(n=xe(e,t,!1))!=null?n:o;return typeof i=="number"?s=>typeof s=="string"?s:i*s:Array.isArray(i)?s=>typeof s=="string"?s:i[s]:typeof i=="function"?i:()=>{}}function lt(e){return he(e,"spacing",8)}function pe(e,t){if(typeof t=="string"||t==null)return t;const o=Math.abs(t),r=e(o);return t>=0?r:typeof r=="number"?-r:`-${r}`}function Ot(e,t){return o=>e.reduce((r,n)=>(r[n]=pe(t,o),r),{})}function It(e,t,o,r){if(t.indexOf(o)===-1)return null;const n=jt(o),i=Ot(n,r),s=e[o];return J(e,s,i)}function it(e,t){const o=lt(e.theme);return Object.keys(e).map(r=>It(e,t,r,o)).reduce(fe,{})}function k(e){return it(e,Pe)}k.propTypes={};k.filterProps=Pe;function A(e){return it(e,Re)}A.propTypes={};A.filterProps=Re;function Pt(e=8){if(e.mui)return e;const t=lt({spacing:e}),o=(...r)=>(r.length===0?[1]:r).map(i=>{const s=t(i);return typeof s=="number"?`${s}px`:s}).join(" ");return o.mui=!0,o}function $e(...e){const t=e.reduce((r,n)=>(n.filterProps.forEach(i=>{r[i]=n}),r),{}),o=r=>Object.keys(r).reduce((n,i)=>t[i]?fe(n,t[i](r)):n,{});return o.propTypes={},o.filterProps=e.reduce((r,n)=>r.concat(n.filterProps),[]),o}function V(e){return typeof e!="number"?e:`${e}px solid`}function U(e,t){return F({prop:e,themeKey:"borders",transform:t})}const Rt=U("border",V),_t=U("borderTop",V),zt=U("borderRight",V),Kt=U("borderBottom",V),Wt=U("borderLeft",V),Mt=U("borderColor"),Lt=U("borderTopColor"),Gt=U("borderRightColor"),Vt=U("borderBottomColor"),Ut=U("borderLeftColor"),Nt=U("outline",V),Yt=U("outlineColor"),ve=e=>{if(e.borderRadius!==void 0&&e.borderRadius!==null){const t=he(e.theme,"shape.borderRadius",4),o=r=>({borderRadius:pe(t,r)});return J(e,e.borderRadius,o)}return null};ve.propTypes={};ve.filterProps=["borderRadius"];$e(Rt,_t,zt,Kt,Wt,Mt,Lt,Gt,Vt,Ut,ve,Nt,Yt);const we=e=>{if(e.gap!==void 0&&e.gap!==null){const t=he(e.theme,"spacing",8),o=r=>({gap:pe(t,r)});return J(e,e.gap,o)}return null};we.propTypes={};we.filterProps=["gap"];const Be=e=>{if(e.columnGap!==void 0&&e.columnGap!==null){const t=he(e.theme,"spacing",8),o=r=>({columnGap:pe(t,r)});return J(e,e.columnGap,o)}return null};Be.propTypes={};Be.filterProps=["columnGap"];const ke=e=>{if(e.rowGap!==void 0&&e.rowGap!==null){const t=he(e.theme,"spacing",8),o=r=>({rowGap:pe(t,r)});return J(e,e.rowGap,o)}return null};ke.propTypes={};ke.filterProps=["rowGap"];const Xt=F({prop:"gridColumn"}),qt=F({prop:"gridRow"}),Qt=F({prop:"gridAutoFlow"}),Jt=F({prop:"gridAutoColumns"}),Zt=F({prop:"gridAutoRows"}),eo=F({prop:"gridTemplateColumns"}),to=F({prop:"gridTemplateRows"}),oo=F({prop:"gridTemplateAreas"}),ro=F({prop:"gridArea"});$e(we,Be,ke,Xt,qt,Qt,Jt,Zt,eo,to,oo,ro);function le(e,t){return t==="grey"?t:e}const no=F({prop:"color",themeKey:"palette",transform:le}),lo=F({prop:"bgcolor",cssProperty:"backgroundColor",themeKey:"palette",transform:le}),io=F({prop:"backgroundColor",themeKey:"palette",transform:le});$e(no,lo,io);function W(e){return e<=1&&e!==0?`${e*100}%`:e}const so=F({prop:"width",transform:W}),_e=e=>{if(e.maxWidth!==void 0&&e.maxWidth!==null){const t=o=>{var r,n;const i=((r=e.theme)==null||(r=r.breakpoints)==null||(r=r.values)==null?void 0:r[o])||Ie[o];return i?((n=e.theme)==null||(n=n.breakpoints)==null?void 0:n.unit)!=="px"?{maxWidth:`${i}${e.theme.breakpoints.unit}`}:{maxWidth:i}:{maxWidth:W(o)}};return J(e,e.maxWidth,t)}return null};_e.filterProps=["maxWidth"];const ao=F({prop:"minWidth",transform:W}),co=F({prop:"height",transform:W}),uo=F({prop:"maxHeight",transform:W}),mo=F({prop:"minHeight",transform:W});F({prop:"size",cssProperty:"width",transform:W});F({prop:"size",cssProperty:"height",transform:W});const fo=F({prop:"boxSizing"});$e(so,_e,ao,co,uo,mo,fo);const st={border:{themeKey:"borders",transform:V},borderTop:{themeKey:"borders",transform:V},borderRight:{themeKey:"borders",transform:V},borderBottom:{themeKey:"borders",transform:V},borderLeft:{themeKey:"borders",transform:V},borderColor:{themeKey:"palette"},borderTopColor:{themeKey:"palette"},borderRightColor:{themeKey:"palette"},borderBottomColor:{themeKey:"palette"},borderLeftColor:{themeKey:"palette"},outline:{themeKey:"borders",transform:V},outlineColor:{themeKey:"palette"},borderRadius:{themeKey:"shape.borderRadius",style:ve},color:{themeKey:"palette",transform:le},bgcolor:{themeKey:"palette",cssProperty:"backgroundColor",transform:le},backgroundColor:{themeKey:"palette",transform:le},p:{style:A},pt:{style:A},pr:{style:A},pb:{style:A},pl:{style:A},px:{style:A},py:{style:A},padding:{style:A},paddingTop:{style:A},paddingRight:{style:A},paddingBottom:{style:A},paddingLeft:{style:A},paddingX:{style:A},paddingY:{style:A},paddingInline:{style:A},paddingInlineStart:{style:A},paddingInlineEnd:{style:A},paddingBlock:{style:A},paddingBlockStart:{style:A},paddingBlockEnd:{style:A},m:{style:k},mt:{style:k},mr:{style:k},mb:{style:k},ml:{style:k},mx:{style:k},my:{style:k},margin:{style:k},marginTop:{style:k},marginRight:{style:k},marginBottom:{style:k},marginLeft:{style:k},marginX:{style:k},marginY:{style:k},marginInline:{style:k},marginInlineStart:{style:k},marginInlineEnd:{style:k},marginBlock:{style:k},marginBlockStart:{style:k},marginBlockEnd:{style:k},displayPrint:{cssProperty:!1,transform:e=>({"@media print":{display:e}})},display:{},overflow:{},textOverflow:{},visibility:{},whiteSpace:{},flexBasis:{},flexDirection:{},flexWrap:{},justifyContent:{},alignItems:{},alignContent:{},order:{},flex:{},flexGrow:{},flexShrink:{},alignSelf:{},justifyItems:{},justifySelf:{},gap:{style:we},rowGap:{style:ke},columnGap:{style:Be},gridColumn:{},gridRow:{},gridAutoFlow:{},gridAutoColumns:{},gridAutoRows:{},gridTemplateColumns:{},gridTemplateRows:{},gridTemplateAreas:{},gridArea:{},position:{},zIndex:{themeKey:"zIndex"},top:{},right:{},bottom:{},left:{},boxShadow:{themeKey:"shadows"},width:{transform:W},maxWidth:{style:_e},minWidth:{transform:W},height:{transform:W},maxHeight:{transform:W},minHeight:{transform:W},boxSizing:{},fontFamily:{themeKey:"typography"},fontSize:{themeKey:"typography"},fontStyle:{themeKey:"typography"},fontWeight:{themeKey:"typography"},letterSpacing:{},textTransform:{},lineHeight:{},textAlign:{},typography:{cssProperty:!1,themeKey:"typography"}};function ho(...e){const t=e.reduce((r,n)=>r.concat(Object.keys(n)),[]),o=new Set(t);return e.every(r=>o.size===Object.keys(r).length)}function po(e,t){return typeof e=="function"?e(t):e}function go(){function e(o,r,n,i){const s={[o]:r,theme:n},c=i[o];if(!c)return{[o]:r};const{cssProperty:d=o,themeKey:b,transform:x,style:w}=c;if(r==null)return null;if(b==="typography"&&r==="inherit")return{[o]:r};const m=xe(n,b)||{};return w?w(s):J(s,r,S=>{let E=Ce(m,x,S);return S===E&&typeof S=="string"&&(E=Ce(m,x,`${o}${S==="default"?"":nt(S)}`,S)),d===!1?E:{[d]:E}})}function t(o){var r;const{sx:n,theme:i={}}=o||{};if(!n)return null;const s=(r=i.unstable_sxConfig)!=null?r:st;function c(d){let b=d;if(typeof d=="function")b=d(i);else if(typeof d!="object")return d;if(!b)return null;const x=Et(i.breakpoints),w=Object.keys(x);let m=x;return Object.keys(b).forEach(y=>{const S=po(b[y],i);if(S!=null)if(typeof S=="object")if(s[y])m=fe(m,e(y,S,i,s));else{const E=J({theme:i},S,g=>({[y]:g}));ho(E,S)?m[y]=t({sx:S,theme:i}):m=fe(m,E)}else m=fe(m,e(y,S,i,s))}),Ft(w,m)}return Array.isArray(n)?n.map(c):c(n)}return t}const at=go();at.filterProps=["sx"];function yo(e,t){const o=this;return o.vars&&typeof o.getColorSchemeSelector=="function"?{[o.getColorSchemeSelector(e).replace(/(\[[^\]]+\])/,"*:where($1)")]:t}:o.palette.mode===e?t:{}}function bo(e){return Object.keys(e).length===0}function So(e=null){const t=B.useContext(rt);return!t||bo(t)?e:t}function Co(e){e=e.slice(1);const t=new RegExp(`.{1,${e.length>=6?2:1}}`,"g");let o=e.match(t);return o&&o[0].length===1&&(o=o.map(r=>r+r)),o?`rgb${o.length===4?"a":""}(${o.map((r,n)=>n<3?parseInt(r,16):Math.round(parseInt(r,16)/255*1e3)/1e3).join(", ")})`:""}function ct(e){if(e.type)return e;if(e.charAt(0)==="#")return ct(Co(e));const t=e.indexOf("("),o=e.substring(0,t);if(["rgb","rgba","hsl","hsla","color"].indexOf(o)===-1)throw new Error(Oe(9,e));let r=e.substring(t+1,e.length-1),n;if(o==="color"){if(r=r.split(" "),n=r.shift(),r.length===4&&r[3].charAt(0)==="/"&&(r[3]=r[3].slice(1)),["srgb","display-p3","a98-rgb","prophoto-rgb","rec-2020"].indexOf(n)===-1)throw new Error(Oe(10,n))}else r=r.split(",");return r=r.map(i=>parseFloat(i)),{type:o,values:r,colorSpace:n}}const te=e=>{const t=ct(e);return t.values.slice(0,3).map((o,r)=>t.type.indexOf("hsl")!==-1&&r!==0?`${o}%`:o).join(" ")},dt=B.createContext(null);function ze(){return B.useContext(dt)}const xo=typeof Symbol=="function"&&Symbol.for,$o=xo?Symbol.for("mui.nested"):"__THEME_NESTED__";function vo(e,t){return typeof t=="function"?t(e):p({},e,t)}function wo(e){const{children:t,theme:o}=e,r=ze(),n=B.useMemo(()=>{const i=r===null?o:vo(r,o);return i!=null&&(i[$o]=r!==null),i},[o,r]);return R.jsx(dt.Provider,{value:n,children:t})}const Bo=["value"],ko=B.createContext();function Ao(e){let{value:t}=e,o=oe(e,Bo);return R.jsx(ko.Provider,p({value:t??!0},o))}const Eo=B.createContext(void 0);function Fo({value:e,children:t}){return R.jsx(Eo.Provider,{value:e,children:t})}const Qe={};function Je(e,t,o,r=!1){return B.useMemo(()=>{const n=e&&t[e]||t;if(typeof o=="function"){const i=o(n),s=e?p({},t,{[e]:i}):i;return r?()=>s:s}return e?p({},t,{[e]:o}):p({},t,o)},[e,t,o,r])}function Do(e){const{children:t,theme:o,themeId:r}=e,n=So(Qe),i=ze()||Qe,s=Je(r,n,o),c=Je(r,i,o,!0),d=s.direction==="rtl";return R.jsx(wo,{theme:c,children:R.jsx(rt.Provider,{value:s,children:R.jsx(Ao,{value:d,children:R.jsx(Fo,{value:s==null?void 0:s.components,children:t})})})})}const Ke="mode",We="color-scheme",ut="data-color-scheme";function Ho(e){const{defaultMode:t="light",defaultLightColorScheme:o="light",defaultDarkColorScheme:r="dark",modeStorageKey:n=Ke,colorSchemeStorageKey:i=We,attribute:s=ut,colorSchemeNode:c="document.documentElement",nonce:d}=e||{};return R.jsx("script",{suppressHydrationWarning:!0,nonce:typeof window>"u"?d:"",dangerouslySetInnerHTML:{__html:`(function() {
+import{j as R,G as Ct,r as B,T as rt}from"./vendor_react-BJ2hNBYs.js";import{_ as oe,a as p,t as xt}from"./vendor_@babel-BRdSvYXJ.js";import{d as Z,c as nt,f as Oe,g as $t}from"./vendor_utils-Ds1aBnGi.js";const vt="$$joy";function wt(e){return e==null||Object.keys(e).length===0}function He(e){const{styles:t,defaultTheme:o={}}=e,r=typeof t=="function"?n=>t(wt(n)?o:n):t;return R.jsx(Ct,{styles:r})}const Bt=["values","unit","step"],kt=e=>{const t=Object.keys(e).map(o=>({key:o,val:e[o]}))||[];return t.sort((o,r)=>o.val-r.val),t.reduce((o,r)=>p({},o,{[r.key]:r.val}),{})};function At(e){const{values:t={xs:0,sm:600,md:900,lg:1200,xl:1536},unit:o="px",step:r=5}=e,n=oe(e,Bt),i=kt(t),s=Object.keys(i);function c(m){return`@media (min-width:${typeof t[m]=="number"?t[m]:m}${o})`}function d(m){return`@media (max-width:${(typeof t[m]=="number"?t[m]:m)-r/100}${o})`}function b(m,y){const S=s.indexOf(y);return`@media (min-width:${typeof t[m]=="number"?t[m]:m}${o}) and (max-width:${(S!==-1&&typeof t[s[S]]=="number"?t[s[S]]:y)-r/100}${o})`}function x(m){return s.indexOf(m)+1<s.length?b(m,s[s.indexOf(m)+1]):c(m)}function w(m){const y=s.indexOf(m);return y===0?c(s[1]):y===s.length-1?d(s[y]):b(m,s[s.indexOf(m)+1]).replace("@media","@media not all and")}return p({keys:s,values:i,up:c,down:d,between:b,only:x,not:w,unit:o},n)}function fe(e,t){return t?Z(e,t,{clone:!1}):e}const Ie={xs:0,sm:600,md:900,lg:1200,xl:1536},Xe={keys:["xs","sm","md","lg","xl"],up:e=>`@media (min-width:${Ie[e]}px)`};function J(e,t,o){const r=e.theme||{};if(Array.isArray(t)){const i=r.breakpoints||Xe;return t.reduce((s,c,d)=>(s[i.up(i.keys[d])]=o(t[d]),s),{})}if(typeof t=="object"){const i=r.breakpoints||Xe;return Object.keys(t).reduce((s,c)=>{if(Object.keys(i.values||Ie).indexOf(c)!==-1){const d=i.up(c);s[d]=o(t[c],c)}else{const d=c;s[d]=t[d]}return s},{})}return o(t)}function Et(e={}){var t;return((t=e.keys)==null?void 0:t.reduce((r,n)=>{const i=e.up(n);return r[i]={},r},{}))||{}}function Ft(e,t){return e.reduce((o,r)=>{const n=o[r];return(!n||Object.keys(n).length===0)&&delete o[r],o},t)}function xe(e,t,o=!0){if(!t||typeof t!="string")return null;if(e&&e.vars&&o){const r=`vars.${t}`.split(".").reduce((n,i)=>n&&n[i]?n[i]:null,e);if(r!=null)return r}return t.split(".").reduce((r,n)=>r&&r[n]!=null?r[n]:null,e)}function Ce(e,t,o,r=o){let n;return typeof e=="function"?n=e(o):Array.isArray(e)?n=e[o]||r:n=xe(e,o)||r,t&&(n=t(n,r,e)),n}function F(e){const{prop:t,cssProperty:o=e.prop,themeKey:r,transform:n}=e,i=s=>{if(s[t]==null)return null;const c=s[t],d=s.theme,b=xe(d,r)||{};return J(s,c,w=>{let m=Ce(b,n,w);return w===m&&typeof w=="string"&&(m=Ce(b,n,`${t}${w==="default"?"":nt(w)}`,w)),o===!1?m:{[o]:m}})};return i.propTypes={},i.filterProps=[t],i}function Dt(e){const t={};return o=>(t[o]===void 0&&(t[o]=e(o)),t[o])}const Ht={m:"margin",p:"padding"},Tt={t:"Top",r:"Right",b:"Bottom",l:"Left",x:["Left","Right"],y:["Top","Bottom"]},qe={marginX:"mx",marginY:"my",paddingX:"px",paddingY:"py"},jt=Dt(e=>{if(e.length>2)if(qe[e])e=qe[e];else return[e];const[t,o]=e.split(""),r=Ht[t],n=Tt[o]||"";return Array.isArray(n)?n.map(i=>r+i):[r+n]}),Pe=["m","mt","mr","mb","ml","mx","my","margin","marginTop","marginRight","marginBottom","marginLeft","marginX","marginY","marginInline","marginInlineStart","marginInlineEnd","marginBlock","marginBlockStart","marginBlockEnd"],Re=["p","pt","pr","pb","pl","px","py","padding","paddingTop","paddingRight","paddingBottom","paddingLeft","paddingX","paddingY","paddingInline","paddingInlineStart","paddingInlineEnd","paddingBlock","paddingBlockStart","paddingBlockEnd"];[...Pe,...Re];function he(e,t,o,r){var n;const i=(n=xe(e,t,!1))!=null?n:o;return typeof i=="number"?s=>typeof s=="string"?s:i*s:Array.isArray(i)?s=>typeof s=="string"?s:i[s]:typeof i=="function"?i:()=>{}}function lt(e){return he(e,"spacing",8)}function pe(e,t){if(typeof t=="string"||t==null)return t;const o=Math.abs(t),r=e(o);return t>=0?r:typeof r=="number"?-r:`-${r}`}function Ot(e,t){return o=>e.reduce((r,n)=>(r[n]=pe(t,o),r),{})}function It(e,t,o,r){if(t.indexOf(o)===-1)return null;const n=jt(o),i=Ot(n,r),s=e[o];return J(e,s,i)}function it(e,t){const o=lt(e.theme);return Object.keys(e).map(r=>It(e,t,r,o)).reduce(fe,{})}function k(e){return it(e,Pe)}k.propTypes={};k.filterProps=Pe;function A(e){return it(e,Re)}A.propTypes={};A.filterProps=Re;function Pt(e=8){if(e.mui)return e;const t=lt({spacing:e}),o=(...r)=>(r.length===0?[1]:r).map(i=>{const s=t(i);return typeof s=="number"?`${s}px`:s}).join(" ");return o.mui=!0,o}function $e(...e){const t=e.reduce((r,n)=>(n.filterProps.forEach(i=>{r[i]=n}),r),{}),o=r=>Object.keys(r).reduce((n,i)=>t[i]?fe(n,t[i](r)):n,{});return o.propTypes={},o.filterProps=e.reduce((r,n)=>r.concat(n.filterProps),[]),o}function V(e){return typeof e!="number"?e:`${e}px solid`}function U(e,t){return F({prop:e,themeKey:"borders",transform:t})}const Rt=U("border",V),_t=U("borderTop",V),zt=U("borderRight",V),Kt=U("borderBottom",V),Wt=U("borderLeft",V),Mt=U("borderColor"),Lt=U("borderTopColor"),Gt=U("borderRightColor"),Vt=U("borderBottomColor"),Ut=U("borderLeftColor"),Nt=U("outline",V),Yt=U("outlineColor"),ve=e=>{if(e.borderRadius!==void 0&&e.borderRadius!==null){const t=he(e.theme,"shape.borderRadius",4),o=r=>({borderRadius:pe(t,r)});return J(e,e.borderRadius,o)}return null};ve.propTypes={};ve.filterProps=["borderRadius"];$e(Rt,_t,zt,Kt,Wt,Mt,Lt,Gt,Vt,Ut,ve,Nt,Yt);const we=e=>{if(e.gap!==void 0&&e.gap!==null){const t=he(e.theme,"spacing",8),o=r=>({gap:pe(t,r)});return J(e,e.gap,o)}return null};we.propTypes={};we.filterProps=["gap"];const Be=e=>{if(e.columnGap!==void 0&&e.columnGap!==null){const t=he(e.theme,"spacing",8),o=r=>({columnGap:pe(t,r)});return J(e,e.columnGap,o)}return null};Be.propTypes={};Be.filterProps=["columnGap"];const ke=e=>{if(e.rowGap!==void 0&&e.rowGap!==null){const t=he(e.theme,"spacing",8),o=r=>({rowGap:pe(t,r)});return J(e,e.rowGap,o)}return null};ke.propTypes={};ke.filterProps=["rowGap"];const Xt=F({prop:"gridColumn"}),qt=F({prop:"gridRow"}),Qt=F({prop:"gridAutoFlow"}),Jt=F({prop:"gridAutoColumns"}),Zt=F({prop:"gridAutoRows"}),eo=F({prop:"gridTemplateColumns"}),to=F({prop:"gridTemplateRows"}),oo=F({prop:"gridTemplateAreas"}),ro=F({prop:"gridArea"});$e(we,Be,ke,Xt,qt,Qt,Jt,Zt,eo,to,oo,ro);function le(e,t){return t==="grey"?t:e}const no=F({prop:"color",themeKey:"palette",transform:le}),lo=F({prop:"bgcolor",cssProperty:"backgroundColor",themeKey:"palette",transform:le}),io=F({prop:"backgroundColor",themeKey:"palette",transform:le});$e(no,lo,io);function W(e){return e<=1&&e!==0?`${e*100}%`:e}const so=F({prop:"width",transform:W}),_e=e=>{if(e.maxWidth!==void 0&&e.maxWidth!==null){const t=o=>{var r,n;const i=((r=e.theme)==null||(r=r.breakpoints)==null||(r=r.values)==null?void 0:r[o])||Ie[o];return i?((n=e.theme)==null||(n=n.breakpoints)==null?void 0:n.unit)!=="px"?{maxWidth:`${i}${e.theme.breakpoints.unit}`}:{maxWidth:i}:{maxWidth:W(o)}};return J(e,e.maxWidth,t)}return null};_e.filterProps=["maxWidth"];const ao=F({prop:"minWidth",transform:W}),co=F({prop:"height",transform:W}),uo=F({prop:"maxHeight",transform:W}),mo=F({prop:"minHeight",transform:W});F({prop:"size",cssProperty:"width",transform:W});F({prop:"size",cssProperty:"height",transform:W});const fo=F({prop:"boxSizing"});$e(so,_e,ao,co,uo,mo,fo);const st={border:{themeKey:"borders",transform:V},borderTop:{themeKey:"borders",transform:V},borderRight:{themeKey:"borders",transform:V},borderBottom:{themeKey:"borders",transform:V},borderLeft:{themeKey:"borders",transform:V},borderColor:{themeKey:"palette"},borderTopColor:{themeKey:"palette"},borderRightColor:{themeKey:"palette"},borderBottomColor:{themeKey:"palette"},borderLeftColor:{themeKey:"palette"},outline:{themeKey:"borders",transform:V},outlineColor:{themeKey:"palette"},borderRadius:{themeKey:"shape.borderRadius",style:ve},color:{themeKey:"palette",transform:le},bgcolor:{themeKey:"palette",cssProperty:"backgroundColor",transform:le},backgroundColor:{themeKey:"palette",transform:le},p:{style:A},pt:{style:A},pr:{style:A},pb:{style:A},pl:{style:A},px:{style:A},py:{style:A},padding:{style:A},paddingTop:{style:A},paddingRight:{style:A},paddingBottom:{style:A},paddingLeft:{style:A},paddingX:{style:A},paddingY:{style:A},paddingInline:{style:A},paddingInlineStart:{style:A},paddingInlineEnd:{style:A},paddingBlock:{style:A},paddingBlockStart:{style:A},paddingBlockEnd:{style:A},m:{style:k},mt:{style:k},mr:{style:k},mb:{style:k},ml:{style:k},mx:{style:k},my:{style:k},margin:{style:k},marginTop:{style:k},marginRight:{style:k},marginBottom:{style:k},marginLeft:{style:k},marginX:{style:k},marginY:{style:k},marginInline:{style:k},marginInlineStart:{style:k},marginInlineEnd:{style:k},marginBlock:{style:k},marginBlockStart:{style:k},marginBlockEnd:{style:k},displayPrint:{cssProperty:!1,transform:e=>({"@media print":{display:e}})},display:{},overflow:{},textOverflow:{},visibility:{},whiteSpace:{},flexBasis:{},flexDirection:{},flexWrap:{},justifyContent:{},alignItems:{},alignContent:{},order:{},flex:{},flexGrow:{},flexShrink:{},alignSelf:{},justifyItems:{},justifySelf:{},gap:{style:we},rowGap:{style:ke},columnGap:{style:Be},gridColumn:{},gridRow:{},gridAutoFlow:{},gridAutoColumns:{},gridAutoRows:{},gridTemplateColumns:{},gridTemplateRows:{},gridTemplateAreas:{},gridArea:{},position:{},zIndex:{themeKey:"zIndex"},top:{},right:{},bottom:{},left:{},boxShadow:{themeKey:"shadows"},width:{transform:W},maxWidth:{style:_e},minWidth:{transform:W},height:{transform:W},maxHeight:{transform:W},minHeight:{transform:W},boxSizing:{},fontFamily:{themeKey:"typography"},fontSize:{themeKey:"typography"},fontStyle:{themeKey:"typography"},fontWeight:{themeKey:"typography"},letterSpacing:{},textTransform:{},lineHeight:{},textAlign:{},typography:{cssProperty:!1,themeKey:"typography"}};function ho(...e){const t=e.reduce((r,n)=>r.concat(Object.keys(n)),[]),o=new Set(t);return e.every(r=>o.size===Object.keys(r).length)}function po(e,t){return typeof e=="function"?e(t):e}function go(){function e(o,r,n,i){const s={[o]:r,theme:n},c=i[o];if(!c)return{[o]:r};const{cssProperty:d=o,themeKey:b,transform:x,style:w}=c;if(r==null)return null;if(b==="typography"&&r==="inherit")return{[o]:r};const m=xe(n,b)||{};return w?w(s):J(s,r,S=>{let E=Ce(m,x,S);return S===E&&typeof S=="string"&&(E=Ce(m,x,`${o}${S==="default"?"":nt(S)}`,S)),d===!1?E:{[d]:E}})}function t(o){var r;const{sx:n,theme:i={}}=o||{};if(!n)return null;const s=(r=i.unstable_sxConfig)!=null?r:st;function c(d){let b=d;if(typeof d=="function")b=d(i);else if(typeof d!="object")return d;if(!b)return null;const x=Et(i.breakpoints),w=Object.keys(x);let m=x;return Object.keys(b).forEach(y=>{const S=po(b[y],i);if(S!=null)if(typeof S=="object")if(s[y])m=fe(m,e(y,S,i,s));else{const E=J({theme:i},S,g=>({[y]:g}));ho(E,S)?m[y]=t({sx:S,theme:i}):m=fe(m,E)}else m=fe(m,e(y,S,i,s))}),Ft(w,m)}return Array.isArray(n)?n.map(c):c(n)}return t}const at=go();at.filterProps=["sx"];function yo(e,t){const o=this;return o.vars&&typeof o.getColorSchemeSelector=="function"?{[o.getColorSchemeSelector(e).replace(/(\[[^\]]+\])/,"*:where($1)")]:t}:o.palette.mode===e?t:{}}function bo(e){return Object.keys(e).length===0}function So(e=null){const t=B.useContext(rt);return!t||bo(t)?e:t}function Co(e){e=e.slice(1);const t=new RegExp(`.{1,${e.length>=6?2:1}}`,"g");let o=e.match(t);return o&&o[0].length===1&&(o=o.map(r=>r+r)),o?`rgb${o.length===4?"a":""}(${o.map((r,n)=>n<3?parseInt(r,16):Math.round(parseInt(r,16)/255*1e3)/1e3).join(", ")})`:""}function ct(e){if(e.type)return e;if(e.charAt(0)==="#")return ct(Co(e));const t=e.indexOf("("),o=e.substring(0,t);if(["rgb","rgba","hsl","hsla","color"].indexOf(o)===-1)throw new Error(Oe(9,e));let r=e.substring(t+1,e.length-1),n;if(o==="color"){if(r=r.split(" "),n=r.shift(),r.length===4&&r[3].charAt(0)==="/"&&(r[3]=r[3].slice(1)),["srgb","display-p3","a98-rgb","prophoto-rgb","rec-2020"].indexOf(n)===-1)throw new Error(Oe(10,n))}else r=r.split(",");return r=r.map(i=>parseFloat(i)),{type:o,values:r,colorSpace:n}}const te=e=>{const t=ct(e);return t.values.slice(0,3).map((o,r)=>t.type.indexOf("hsl")!==-1&&r!==0?`${o}%`:o).join(" ")},dt=B.createContext(null);function ze(){return B.useContext(dt)}const xo=typeof Symbol=="function"&&Symbol.for,$o=xo?Symbol.for("mui.nested"):"__THEME_NESTED__";function vo(e,t){return typeof t=="function"?t(e):p({},e,t)}function wo(e){const{children:t,theme:o}=e,r=ze(),n=B.useMemo(()=>{const i=r===null?o:vo(r,o);return i!=null&&(i[$o]=r!==null),i},[o,r]);return R.jsx(dt.Provider,{value:n,children:t})}const Bo=["value"],ko=B.createContext();function Ao(e){let{value:t}=e,o=oe(e,Bo);return R.jsx(ko.Provider,p({value:t??!0},o))}const Eo=B.createContext(void 0);function Fo({value:e,children:t}){return R.jsx(Eo.Provider,{value:e,children:t})}const Qe={};function Je(e,t,o,r=!1){return B.useMemo(()=>{const n=e&&t[e]||t;if(typeof o=="function"){const i=o(n),s=e?p({},t,{[e]:i}):i;return r?()=>s:s}return e?p({},t,{[e]:o}):p({},t,o)},[e,t,o,r])}function Do(e){const{children:t,theme:o,themeId:r}=e,n=So(Qe),i=ze()||Qe,s=Je(r,n,o),c=Je(r,i,o,!0),d=s.direction==="rtl";return R.jsx(wo,{theme:c,children:R.jsx(rt.Provider,{value:s,children:R.jsx(Ao,{value:d,children:R.jsx(Fo,{value:s==null?void 0:s.components,children:t})})})})}const Ke="mode",We="color-scheme",ut="data-color-scheme";function Ho(e){const{defaultMode:t="light",defaultLightColorScheme:o="light",defaultDarkColorScheme:r="dark",modeStorageKey:n=Ke,colorSchemeStorageKey:i=We,attribute:s=ut,colorSchemeNode:c="document.documentElement",nonce:d}=e||{};return R.jsx("script",{suppressHydrationWarning:!0,nonce:typeof window>"u"?d:"",dangerouslySetInnerHTML:{__html:`(function() {
 try {
   var mode = localStorage.getItem('${n}') || '${t}';
   var colorScheme = '';
