+++ assets/vendor_material-GI4Anplu.js
@@ -1,4 +1,4 @@
-var Nr=Object.defineProperty;var Dr=(e,t,o)=>t in e?Nr(e,t,{enumerable:!0,configurable:!0,writable:!0,value:o}):e[t]=o;var dt=(e,t,o)=>Dr(e,typeof t!="symbol"?t+"":t,o);import{g as jr,l as tr,d as or,c as Wr,a as rr,s as io,p as Ur,b as oe,e as re,f as wt,h as Rt,i as mt,j as Hr,k as Vr,m as Kr,n as qr,o as Gr,q as _r,u as Xr,r as Yr,t as Zr,T as Jr,v as Qr,G as en,w as tn,x as on,y as rn,z as Oe,A as nn,B as an,C as sn,D as nr}from"./vendor_system-DdpyrVMS.js";import{r as c,a as ir,j as w,R as ln,b as cn,c as ar,k as je,d as dn,e as xt}from"./vendor_react-BO86OV0u.js";import{c as U}from"./vendor_clsx-B-dksMZM.js";function Ve(e,...t){const o=new URL(`https://mui.com/production-error/?code=${e}`);return t.forEach(r=>o.searchParams.append("args[]",r)),`Minified MUI error #${e}; visit ${o} for the full message.`}const Ie="$$material";function De(e){if(typeof e!="object"||e===null)return!1;const t=Object.getPrototypeOf(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)}function sr(e){if(c.isValidElement(e)||ir.isValidElementType(e)||!De(e))return e;const t={};return Object.keys(e).forEach(o=>{t[o]=sr(e[o])}),t}function Ae(e,t,o={clone:!0}){const r=o.clone?{...e}:e;return De(e)&&De(t)&&Object.keys(t).forEach(n=>{c.isValidElement(t[n])||ir.isValidElementType(t[n])?r[n]=t[n]:De(t[n])&&Object.prototype.hasOwnProperty.call(e,n)&&De(e[n])?r[n]=Ae(e[n],t[n],o):o.clone?r[n]=De(t[n])?sr(t[n]):t[n]:r[n]=t[n]}),r}const vt={black:"#000",white:"#fff"},un={50:"#fafafa",100:"#f5f5f5",200:"#eeeeee",300:"#e0e0e0",400:"#bdbdbd",500:"#9e9e9e",600:"#757575",700:"#616161",800:"#424242",900:"#212121",A100:"#f5f5f5",A200:"#eeeeee",A400:"#bdbdbd",A700:"#616161"},Je={50:"#f3e5f5",100:"#e1bee7",200:"#ce93d8",300:"#ba68c8",400:"#ab47bc",500:"#9c27b0",600:"#8e24aa",700:"#7b1fa2",800:"#6a1b9a",900:"#4a148c",A100:"#ea80fc",A200:"#e040fb",A400:"#d500f9",A700:"#aa00ff"},Qe={50:"#ffebee",100:"#ffcdd2",200:"#ef9a9a",300:"#e57373",400:"#ef5350",500:"#f44336",600:"#e53935",700:"#d32f2f",800:"#c62828",900:"#b71c1c",A100:"#ff8a80",A200:"#ff5252",A400:"#ff1744",A700:"#d50000"},ut={50:"#fff3e0",100:"#ffe0b2",200:"#ffcc80",300:"#ffb74d",400:"#ffa726",500:"#ff9800",600:"#fb8c00",700:"#f57c00",800:"#ef6c00",900:"#e65100",A100:"#ffd180",A200:"#ffab40",A400:"#ff9100",A700:"#ff6d00"},et={50:"#e3f2fd",100:"#bbdefb",200:"#90caf9",300:"#64b5f6",400:"#42a5f5",500:"#2196f3",600:"#1e88e5",700:"#1976d2",800:"#1565c0",900:"#0d47a1",A100:"#82b1ff",A200:"#448aff",A400:"#2979ff",A700:"#2962ff"},tt={50:"#e1f5fe",100:"#b3e5fc",200:"#81d4fa",300:"#4fc3f7",400:"#29b6f6",500:"#03a9f4",600:"#039be5",700:"#0288d1",800:"#0277bd",900:"#01579b",A100:"#80d8ff",A200:"#40c4ff",A400:"#00b0ff",A700:"#0091ea"},ot={50:"#e8f5e9",100:"#c8e6c9",200:"#a5d6a7",300:"#81c784",400:"#66bb6a",500:"#4caf50",600:"#43a047",700:"#388e3c",800:"#2e7d32",900:"#1b5e20",A100:"#b9f6ca",A200:"#69f0ae",A400:"#00e676",A700:"#00c853"};function lr(){return{text:{primary:"rgba(0, 0, 0, 0.87)",secondary:"rgba(0, 0, 0, 0.6)",disabled:"rgba(0, 0, 0, 0.38)"},divider:"rgba(0, 0, 0, 0.12)",background:{paper:vt.white,default:vt.white},action:{active:"rgba(0, 0, 0, 0.54)",hover:"rgba(0, 0, 0, 0.04)",hoverOpacity:.04,selected:"rgba(0, 0, 0, 0.08)",selectedOpacity:.08,disabled:"rgba(0, 0, 0, 0.26)",disabledBackground:"rgba(0, 0, 0, 0.12)",disabledOpacity:.38,focus:"rgba(0, 0, 0, 0.12)",focusOpacity:.12,activatedOpacity:.12}}}const pn=lr();function cr(){return{text:{primary:vt.white,secondary:"rgba(255, 255, 255, 0.7)",disabled:"rgba(255, 255, 255, 0.5)",icon:"rgba(255, 255, 255, 0.5)"},divider:"rgba(255, 255, 255, 0.12)",background:{paper:"#121212",default:"#121212"},action:{active:vt.white,hover:"rgba(255, 255, 255, 0.08)",hoverOpacity:.08,selected:"rgba(255, 255, 255, 0.16)",selectedOpacity:.16,disabled:"rgba(255, 255, 255, 0.3)",disabledBackground:"rgba(255, 255, 255, 0.12)",disabledOpacity:.38,focus:"rgba(255, 255, 255, 0.12)",focusOpacity:.12,activatedOpacity:.24}}}const So=cr();function Co(e,t,o,r){const n=r.light||r,a=r.dark||r*1.5;e[t]||(e.hasOwnProperty(o)?e[t]=e[o]:t==="light"?e.light=tr(e.main,n):t==="dark"&&(e.dark=or(e.main,a)))}function fn(e="light"){return e==="dark"?{main:et[200],light:et[50],dark:et[400]}:{main:et[700],light:et[400],dark:et[800]}}function gn(e="light"){return e==="dark"?{main:Je[200],light:Je[50],dark:Je[400]}:{main:Je[500],light:Je[300],dark:Je[700]}}function mn(e="light"){return e==="dark"?{main:Qe[500],light:Qe[300],dark:Qe[700]}:{main:Qe[700],light:Qe[400],dark:Qe[800]}}function hn(e="light"){return e==="dark"?{main:tt[400],light:tt[300],dark:tt[700]}:{main:tt[700],light:tt[500],dark:tt[900]}}function bn(e="light"){return e==="dark"?{main:ot[400],light:ot[300],dark:ot[700]}:{main:ot[800],light:ot[500],dark:ot[900]}}function yn(e="light"){return e==="dark"?{main:ut[400],light:ut[300],dark:ut[700]}:{main:"#ed6c02",light:ut[500],dark:ut[900]}}function ao(e){const{mode:t="light",contrastThreshold:o=3,tonalOffset:r=.2,...n}=e,a=e.primary||fn(t),s=e.secondary||gn(t),l=e.error||mn(t),u=e.info||hn(t),d=e.success||bn(t),b=e.warning||yn(t);function p(v){return jr(v,So.text.primary)>=o?So.text.primary:pn.text.primary}const y=({color:v,name:C,mainShade:S=500,lightShade:R=300,darkShade:h=700})=>{if(v={...v},!v.main&&v[S]&&(v.main=v[S]),!v.hasOwnProperty("main"))throw new Error(Ve(11,C?` (${C})`:"",S));if(typeof v.main!="string")throw new Error(Ve(12,C?` (${C})`:"",JSON.stringify(v.main)));return Co(v,"light",R,r),Co(v,"dark",h,r),v.contrastText||(v.contrastText=p(v.main)),v};let m;return t==="light"?m=lr():t==="dark"&&(m=cr()),Ae({common:{...vt},mode:t,primary:y({color:a,name:"primary"}),secondary:y({color:s,name:"secondary",mainShade:"A400",lightShade:"A200",darkShade:"A700"}),error:y({color:l,name:"error"}),warning:y({color:b,name:"warning"}),info:y({color:u,name:"info"}),success:y({color:d,name:"success"}),grey:un,contrastThreshold:o,getContrastText:p,augmentColor:y,tonalOffset:r,...m},n)}const wo=e=>e,vn=()=>{let e=wo;return{configure(t){e=t},generate(t){return e(t)},reset(){e=wo}}},dr=vn(),xn={active:"active",checked:"checked",completed:"completed",disabled:"disabled",error:"error",expanded:"expanded",focused:"focused",focusVisible:"focusVisible",open:"open",readOnly:"readOnly",required:"required",selected:"selected"};function Z(e,t,o="Mui"){const r=xn[t];return r?`${o}-${r}`:`${dr.generate(e)}-${t}`}function Sn(e,t){return{toolbar:{minHeight:56,[e.up("xs")]:{"@media (orientation: landscape)":{minHeight:48}},[e.up("sm")]:{minHeight:64}},...t}}function Cn(e){return Math.round(e*1e5)/1e5}const Ro={textTransform:"uppercase"},ko='"Roboto", "Helvetica", "Arial", sans-serif';function ur(e,t){const{fontFamily:o=ko,fontSize:r=14,fontWeightLight:n=300,fontWeightRegular:a=400,fontWeightMedium:s=500,fontWeightBold:l=700,htmlFontSize:u=16,allVariants:d,pxToRem:b,...p}=typeof t=="function"?t(e):t,y=r/14,m=b||(C=>`${C/u*y}rem`),g=(C,S,R,h,f)=>({fontFamily:o,fontWeight:C,fontSize:m(S),lineHeight:R,...o===ko?{letterSpacing:`${Cn(h/S)}em`}:{},...f,...d}),v={h1:g(n,96,1.167,-1.5),h2:g(n,60,1.2,-.5),h3:g(a,48,1.167,0),h4:g(a,34,1.235,.25),h5:g(a,24,1.334,0),h6:g(s,20,1.6,.15),subtitle1:g(a,16,1.75,.15),subtitle2:g(s,14,1.57,.1),body1:g(a,16,1.5,.15),body2:g(a,14,1.43,.15),button:g(s,14,1.75,.4,Ro),caption:g(a,12,1.66,.4),overline:g(a,12,2.66,1,Ro),inherit:{fontFamily:"inherit",fontWeight:"inherit",fontSize:"inherit",lineHeight:"inherit",letterSpacing:"inherit"}};return Ae({htmlFontSize:u,pxToRem:m,fontFamily:o,fontSize:r,fontWeightLight:n,fontWeightRegular:a,fontWeightMedium:s,fontWeightBold:l,...v},p,{clone:!1})}const wn=.2,Rn=.14,kn=.12;function ce(...e){return[`${e[0]}px ${e[1]}px ${e[2]}px ${e[3]}px rgba(0,0,0,${wn})`,`${e[4]}px ${e[5]}px ${e[6]}px ${e[7]}px rgba(0,0,0,${Rn})`,`${e[8]}px ${e[9]}px ${e[10]}px ${e[11]}px rgba(0,0,0,${kn})`].join(",")}const Pn=["none",ce(0,2,1,-1,0,1,1,0,0,1,3,0),ce(0,3,1,-2,0,2,2,0,0,1,5,0),ce(0,3,3,-2,0,3,4,0,0,1,8,0),ce(0,2,4,-1,0,4,5,0,0,1,10,0),ce(0,3,5,-1,0,5,8,0,0,1,14,0),ce(0,3,5,-1,0,6,10,0,0,1,18,0),ce(0,4,5,-2,0,7,10,1,0,2,16,1),ce(0,5,5,-3,0,8,10,1,0,3,14,2),ce(0,5,6,-3,0,9,12,1,0,3,16,2),ce(0,6,6,-3,0,10,14,1,0,4,18,3),ce(0,6,7,-4,0,11,15,1,0,4,20,3),ce(0,7,8,-4,0,12,17,2,0,5,22,4),ce(0,7,8,-4,0,13,19,2,0,5,24,4),ce(0,7,9,-4,0,14,21,2,0,5,26,4),ce(0,8,9,-5,0,15,22,2,0,6,28,5),ce(0,8,10,-5,0,16,24,2,0,6,30,5),ce(0,8,11,-5,0,17,26,2,0,6,32,5),ce(0,9,11,-5,0,18,28,2,0,7,34,6),ce(0,9,12,-6,0,19,29,2,0,7,36,6),ce(0,10,13,-6,0,20,31,3,0,8,38,7),ce(0,10,13,-6,0,21,33,3,0,8,40,7),ce(0,10,14,-6,0,22,35,3,0,8,42,7),ce(0,11,14,-7,0,23,36,3,0,9,44,8),ce(0,11,15,-7,0,24,38,3,0,9,46,8)],In={easeInOut:"cubic-bezier(0.4, 0, 0.2, 1)",easeOut:"cubic-bezier(0.0, 0, 0.2, 1)",easeIn:"cubic-bezier(0.4, 0, 1, 1)",sharp:"cubic-bezier(0.4, 0, 0.6, 1)"},Tn={shortest:150,shorter:200,short:250,standard:300,complex:375,enteringScreen:225,leavingScreen:195};function Po(e){return`${Math.round(e)}ms`}function $n(e){if(!e)return 0;const t=e/36;return Math.min(Math.round((4+15*t**.25+t/5)*10),3e3)}function Mn(e){const t={...In,...e.easing},o={...Tn,...e.duration};return{getAutoHeightDuration:$n,create:(n=["all"],a={})=>{const{duration:s=o.standard,easing:l=t.easeInOut,delay:u=0,...d}=a;return(Array.isArray(n)?n:[n]).map(b=>`${b} ${typeof s=="string"?s:Po(s)} ${l} ${typeof u=="string"?u:Po(u)}`).join(",")},...e,easing:t,duration:o}}const En={mobileStepper:1e3,fab:1050,speedDial:1050,appBar:1100,drawer:1200,modal:1300,snackbar:1400,tooltip:1500};function Bn(e){return De(e)||typeof e>"u"||typeof e=="string"||typeof e=="boolean"||typeof e=="number"||Array.isArray(e)}function pr(e={}){const t={...e};function o(r){const n=Object.entries(r);for(let a=0;a<n.length;a++){const[s,l]=n[a];!Bn(l)||s.startsWith("unstable_")?delete r[s]:De(l)&&(r[s]={...l},o(r[s]))}}return o(t),`import { unstable_createBreakpoints as createBreakpoints, createTransitions } from '@mui/material/styles';
+var Nr=Object.defineProperty;var Dr=(e,t,o)=>t in e?Nr(e,t,{enumerable:!0,configurable:!0,writable:!0,value:o}):e[t]=o;var dt=(e,t,o)=>Dr(e,typeof t!="symbol"?t+"":t,o);import{g as jr,l as tr,d as or,c as Wr,a as rr,s as io,p as Ur,b as oe,e as re,f as wt,h as Rt,i as mt,j as Hr,k as Vr,m as Kr,n as qr,o as Gr,q as _r,u as Xr,r as Yr,t as Zr,T as Jr,v as Qr,G as en,w as tn,x as on,y as rn,z as Oe,A as nn,B as an,C as sn,D as nr}from"./vendor_system-bhe-e0RG.js";import{r as c,a as ir,j as w,R as ln,b as cn,c as ar,k as je,d as dn,e as xt}from"./vendor_react-BJ2hNBYs.js";import{c as U}from"./vendor_clsx-B-dksMZM.js";function Ve(e,...t){const o=new URL(`https://mui.com/production-error/?code=${e}`);return t.forEach(r=>o.searchParams.append("args[]",r)),`Minified MUI error #${e}; visit ${o} for the full message.`}const Ie="$$material";function De(e){if(typeof e!="object"||e===null)return!1;const t=Object.getPrototypeOf(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)}function sr(e){if(c.isValidElement(e)||ir.isValidElementType(e)||!De(e))return e;const t={};return Object.keys(e).forEach(o=>{t[o]=sr(e[o])}),t}function Ae(e,t,o={clone:!0}){const r=o.clone?{...e}:e;return De(e)&&De(t)&&Object.keys(t).forEach(n=>{c.isValidElement(t[n])||ir.isValidElementType(t[n])?r[n]=t[n]:De(t[n])&&Object.prototype.hasOwnProperty.call(e,n)&&De(e[n])?r[n]=Ae(e[n],t[n],o):o.clone?r[n]=De(t[n])?sr(t[n]):t[n]:r[n]=t[n]}),r}const vt={black:"#000",white:"#fff"},un={50:"#fafafa",100:"#f5f5f5",200:"#eeeeee",300:"#e0e0e0",400:"#bdbdbd",500:"#9e9e9e",600:"#757575",700:"#616161",800:"#424242",900:"#212121",A100:"#f5f5f5",A200:"#eeeeee",A400:"#bdbdbd",A700:"#616161"},Je={50:"#f3e5f5",100:"#e1bee7",200:"#ce93d8",300:"#ba68c8",400:"#ab47bc",500:"#9c27b0",600:"#8e24aa",700:"#7b1fa2",800:"#6a1b9a",900:"#4a148c",A100:"#ea80fc",A200:"#e040fb",A400:"#d500f9",A700:"#aa00ff"},Qe={50:"#ffebee",100:"#ffcdd2",200:"#ef9a9a",300:"#e57373",400:"#ef5350",500:"#f44336",600:"#e53935",700:"#d32f2f",800:"#c62828",900:"#b71c1c",A100:"#ff8a80",A200:"#ff5252",A400:"#ff1744",A700:"#d50000"},ut={50:"#fff3e0",100:"#ffe0b2",200:"#ffcc80",300:"#ffb74d",400:"#ffa726",500:"#ff9800",600:"#fb8c00",700:"#f57c00",800:"#ef6c00",900:"#e65100",A100:"#ffd180",A200:"#ffab40",A400:"#ff9100",A700:"#ff6d00"},et={50:"#e3f2fd",100:"#bbdefb",200:"#90caf9",300:"#64b5f6",400:"#42a5f5",500:"#2196f3",600:"#1e88e5",700:"#1976d2",800:"#1565c0",900:"#0d47a1",A100:"#82b1ff",A200:"#448aff",A400:"#2979ff",A700:"#2962ff"},tt={50:"#e1f5fe",100:"#b3e5fc",200:"#81d4fa",300:"#4fc3f7",400:"#29b6f6",500:"#03a9f4",600:"#039be5",700:"#0288d1",800:"#0277bd",900:"#01579b",A100:"#80d8ff",A200:"#40c4ff",A400:"#00b0ff",A700:"#0091ea"},ot={50:"#e8f5e9",100:"#c8e6c9",200:"#a5d6a7",300:"#81c784",400:"#66bb6a",500:"#4caf50",600:"#43a047",700:"#388e3c",800:"#2e7d32",900:"#1b5e20",A100:"#b9f6ca",A200:"#69f0ae",A400:"#00e676",A700:"#00c853"};function lr(){return{text:{primary:"rgba(0, 0, 0, 0.87)",secondary:"rgba(0, 0, 0, 0.6)",disabled:"rgba(0, 0, 0, 0.38)"},divider:"rgba(0, 0, 0, 0.12)",background:{paper:vt.white,default:vt.white},action:{active:"rgba(0, 0, 0, 0.54)",hover:"rgba(0, 0, 0, 0.04)",hoverOpacity:.04,selected:"rgba(0, 0, 0, 0.08)",selectedOpacity:.08,disabled:"rgba(0, 0, 0, 0.26)",disabledBackground:"rgba(0, 0, 0, 0.12)",disabledOpacity:.38,focus:"rgba(0, 0, 0, 0.12)",focusOpacity:.12,activatedOpacity:.12}}}const pn=lr();function cr(){return{text:{primary:vt.white,secondary:"rgba(255, 255, 255, 0.7)",disabled:"rgba(255, 255, 255, 0.5)",icon:"rgba(255, 255, 255, 0.5)"},divider:"rgba(255, 255, 255, 0.12)",background:{paper:"#121212",default:"#121212"},action:{active:vt.white,hover:"rgba(255, 255, 255, 0.08)",hoverOpacity:.08,selected:"rgba(255, 255, 255, 0.16)",selectedOpacity:.16,disabled:"rgba(255, 255, 255, 0.3)",disabledBackground:"rgba(255, 255, 255, 0.12)",disabledOpacity:.38,focus:"rgba(255, 255, 255, 0.12)",focusOpacity:.12,activatedOpacity:.24}}}const So=cr();function Co(e,t,o,r){const n=r.light||r,a=r.dark||r*1.5;e[t]||(e.hasOwnProperty(o)?e[t]=e[o]:t==="light"?e.light=tr(e.main,n):t==="dark"&&(e.dark=or(e.main,a)))}function fn(e="light"){return e==="dark"?{main:et[200],light:et[50],dark:et[400]}:{main:et[700],light:et[400],dark:et[800]}}function gn(e="light"){return e==="dark"?{main:Je[200],light:Je[50],dark:Je[400]}:{main:Je[500],light:Je[300],dark:Je[700]}}function mn(e="light"){return e==="dark"?{main:Qe[500],light:Qe[300],dark:Qe[700]}:{main:Qe[700],light:Qe[400],dark:Qe[800]}}function hn(e="light"){return e==="dark"?{main:tt[400],light:tt[300],dark:tt[700]}:{main:tt[700],light:tt[500],dark:tt[900]}}function bn(e="light"){return e==="dark"?{main:ot[400],light:ot[300],dark:ot[700]}:{main:ot[800],light:ot[500],dark:ot[900]}}function yn(e="light"){return e==="dark"?{main:ut[400],light:ut[300],dark:ut[700]}:{main:"#ed6c02",light:ut[500],dark:ut[900]}}function ao(e){const{mode:t="light",contrastThreshold:o=3,tonalOffset:r=.2,...n}=e,a=e.primary||fn(t),s=e.secondary||gn(t),l=e.error||mn(t),u=e.info||hn(t),d=e.success||bn(t),b=e.warning||yn(t);function p(v){return jr(v,So.text.primary)>=o?So.text.primary:pn.text.primary}const y=({color:v,name:C,mainShade:S=500,lightShade:R=300,darkShade:h=700})=>{if(v={...v},!v.main&&v[S]&&(v.main=v[S]),!v.hasOwnProperty("main"))throw new Error(Ve(11,C?` (${C})`:"",S));if(typeof v.main!="string")throw new Error(Ve(12,C?` (${C})`:"",JSON.stringify(v.main)));return Co(v,"light",R,r),Co(v,"dark",h,r),v.contrastText||(v.contrastText=p(v.main)),v};let m;return t==="light"?m=lr():t==="dark"&&(m=cr()),Ae({common:{...vt},mode:t,primary:y({color:a,name:"primary"}),secondary:y({color:s,name:"secondary",mainShade:"A400",lightShade:"A200",darkShade:"A700"}),error:y({color:l,name:"error"}),warning:y({color:b,name:"warning"}),info:y({color:u,name:"info"}),success:y({color:d,name:"success"}),grey:un,contrastThreshold:o,getContrastText:p,augmentColor:y,tonalOffset:r,...m},n)}const wo=e=>e,vn=()=>{let e=wo;return{configure(t){e=t},generate(t){return e(t)},reset(){e=wo}}},dr=vn(),xn={active:"active",checked:"checked",completed:"completed",disabled:"disabled",error:"error",expanded:"expanded",focused:"focused",focusVisible:"focusVisible",open:"open",readOnly:"readOnly",required:"required",selected:"selected"};function Z(e,t,o="Mui"){const r=xn[t];return r?`${o}-${r}`:`${dr.generate(e)}-${t}`}function Sn(e,t){return{toolbar:{minHeight:56,[e.up("xs")]:{"@media (orientation: landscape)":{minHeight:48}},[e.up("sm")]:{minHeight:64}},...t}}function Cn(e){return Math.round(e*1e5)/1e5}const Ro={textTransform:"uppercase"},ko='"Roboto", "Helvetica", "Arial", sans-serif';function ur(e,t){const{fontFamily:o=ko,fontSize:r=14,fontWeightLight:n=300,fontWeightRegular:a=400,fontWeightMedium:s=500,fontWeightBold:l=700,htmlFontSize:u=16,allVariants:d,pxToRem:b,...p}=typeof t=="function"?t(e):t,y=r/14,m=b||(C=>`${C/u*y}rem`),g=(C,S,R,h,f)=>({fontFamily:o,fontWeight:C,fontSize:m(S),lineHeight:R,...o===ko?{letterSpacing:`${Cn(h/S)}em`}:{},...f,...d}),v={h1:g(n,96,1.167,-1.5),h2:g(n,60,1.2,-.5),h3:g(a,48,1.167,0),h4:g(a,34,1.235,.25),h5:g(a,24,1.334,0),h6:g(s,20,1.6,.15),subtitle1:g(a,16,1.75,.15),subtitle2:g(s,14,1.57,.1),body1:g(a,16,1.5,.15),body2:g(a,14,1.43,.15),button:g(s,14,1.75,.4,Ro),caption:g(a,12,1.66,.4),overline:g(a,12,2.66,1,Ro),inherit:{fontFamily:"inherit",fontWeight:"inherit",fontSize:"inherit",lineHeight:"inherit",letterSpacing:"inherit"}};return Ae({htmlFontSize:u,pxToRem:m,fontFamily:o,fontSize:r,fontWeightLight:n,fontWeightRegular:a,fontWeightMedium:s,fontWeightBold:l,...v},p,{clone:!1})}const wn=.2,Rn=.14,kn=.12;function ce(...e){return[`${e[0]}px ${e[1]}px ${e[2]}px ${e[3]}px rgba(0,0,0,${wn})`,`${e[4]}px ${e[5]}px ${e[6]}px ${e[7]}px rgba(0,0,0,${Rn})`,`${e[8]}px ${e[9]}px ${e[10]}px ${e[11]}px rgba(0,0,0,${kn})`].join(",")}const Pn=["none",ce(0,2,1,-1,0,1,1,0,0,1,3,0),ce(0,3,1,-2,0,2,2,0,0,1,5,0),ce(0,3,3,-2,0,3,4,0,0,1,8,0),ce(0,2,4,-1,0,4,5,0,0,1,10,0),ce(0,3,5,-1,0,5,8,0,0,1,14,0),ce(0,3,5,-1,0,6,10,0,0,1,18,0),ce(0,4,5,-2,0,7,10,1,0,2,16,1),ce(0,5,5,-3,0,8,10,1,0,3,14,2),ce(0,5,6,-3,0,9,12,1,0,3,16,2),ce(0,6,6,-3,0,10,14,1,0,4,18,3),ce(0,6,7,-4,0,11,15,1,0,4,20,3),ce(0,7,8,-4,0,12,17,2,0,5,22,4),ce(0,7,8,-4,0,13,19,2,0,5,24,4),ce(0,7,9,-4,0,14,21,2,0,5,26,4),ce(0,8,9,-5,0,15,22,2,0,6,28,5),ce(0,8,10,-5,0,16,24,2,0,6,30,5),ce(0,8,11,-5,0,17,26,2,0,6,32,5),ce(0,9,11,-5,0,18,28,2,0,7,34,6),ce(0,9,12,-6,0,19,29,2,0,7,36,6),ce(0,10,13,-6,0,20,31,3,0,8,38,7),ce(0,10,13,-6,0,21,33,3,0,8,40,7),ce(0,10,14,-6,0,22,35,3,0,8,42,7),ce(0,11,14,-7,0,23,36,3,0,9,44,8),ce(0,11,15,-7,0,24,38,3,0,9,46,8)],In={easeInOut:"cubic-bezier(0.4, 0, 0.2, 1)",easeOut:"cubic-bezier(0.0, 0, 0.2, 1)",easeIn:"cubic-bezier(0.4, 0, 1, 1)",sharp:"cubic-bezier(0.4, 0, 0.6, 1)"},Tn={shortest:150,shorter:200,short:250,standard:300,complex:375,enteringScreen:225,leavingScreen:195};function Po(e){return`${Math.round(e)}ms`}function $n(e){if(!e)return 0;const t=e/36;return Math.min(Math.round((4+15*t**.25+t/5)*10),3e3)}function Mn(e){const t={...In,...e.easing},o={...Tn,...e.duration};return{getAutoHeightDuration:$n,create:(n=["all"],a={})=>{const{duration:s=o.standard,easing:l=t.easeInOut,delay:u=0,...d}=a;return(Array.isArray(n)?n:[n]).map(b=>`${b} ${typeof s=="string"?s:Po(s)} ${l} ${typeof u=="string"?u:Po(u)}`).join(",")},...e,easing:t,duration:o}}const En={mobileStepper:1e3,fab:1050,speedDial:1050,appBar:1100,drawer:1200,modal:1300,snackbar:1400,tooltip:1500};function Bn(e){return De(e)||typeof e>"u"||typeof e=="string"||typeof e=="boolean"||typeof e=="number"||Array.isArray(e)}function pr(e={}){const t={...e};function o(r){const n=Object.entries(r);for(let a=0;a<n.length;a++){const[s,l]=n[a];!Bn(l)||s.startsWith("unstable_")?delete r[s]:De(l)&&(r[s]={...l},o(r[s]))}}return o(t),`import { unstable_createBreakpoints as createBreakpoints, createTransitions } from '@mui/material/styles';
 
 const theme = ${JSON.stringify(t,null,2)};
