+++ assets/vendor_private-theming-Cy3868wu.js
@@ -1,0 +1,1 @@
+import{r as u,j as c}from"./vendor_react-BJ2hNBYs.js";const s=u.createContext(null);function i(){return u.useContext(s)}const l=typeof Symbol=="function"&&Symbol.for,f=l?Symbol.for("mui.nested"):"__THEME_NESTED__";function h(e,t){return typeof t=="function"?t(e):{...e,...t}}function x(e){const{children:t,theme:n}=e,o=i(),m=u.useMemo(()=>{const r=o===null?{...n}:h(o,n);return r!=null&&(r[f]=o!==null),r},[n,o]);return c.jsx(s.Provider,{value:m,children:t})}export{x as T,i as u};
