--- assets/vendor_swr-Dyr5jNOL.js
@@ -1,1 +1,0 @@
-import{f as j,r as C}from"./vendor_react-BO86OV0u.js";import"./vendor_use-sync-external-store-D1USJxqC.js";import{d as ct}from"./vendor_dequal-CnqicUFC.js";const R=new WeakMap,T=()=>{},l=T(),I=Object,p=t=>t===l,h=t=>typeof t=="function",S=(t,e)=>({...t,...e}),rt=t=>h(t.then),L={},y={},k="undefined",U=typeof window!=k,F=typeof document!=k,at=(t,e)=>{const o=R.get(t);return[()=>!p(e)&&t.get(e)||L,c=>{if(!p(e)){const i=t.get(e);e in y||(y[e]=i),o[5](e,S(i,c),i||L)}},o[6],()=>!p(e)&&e in y?y[e]:!p(e)&&t.get(e)||L]};let V=!0;const ut=()=>V,[W,P]=U&&window.addEventListener?[window.addEventListener.bind(window),window.removeEventListener.bind(window)]:[T,T],ft=()=>{const t=F&&document.visibilityState;return p(t)||t!=="hidden"},lt=t=>(F&&document.addEventListener("visibilitychange",t),W("focus",t),()=>{F&&document.removeEventListener("visibilitychange",t),P("focus",t)}),dt=t=>{const e=()=>{V=!0,t()},o=()=>{V=!1};return W("online",e),W("offline",o),()=>{P("online",e),P("offline",o)}},Et={isOnline:ut,isVisible:ft},pt={initFocus:lt,initReconnect:dt};j.useId;const $=!U||"Deno"in globalThis,mt=$?C.useEffect:C.useLayoutEffect,N=typeof navigator<"u"&&navigator.connection,X=!$&&N&&(["slow-2g","2g"].includes(N.effectiveType)||N.saveData),b=new WeakMap,x=(t,e)=>I.prototype.toString.call(t)===`[object ${e}]`;let wt=0;const M=t=>{const e=typeof t,o=x(t,"Date"),c=x(t,"RegExp"),i=x(t,"Object");let n,s;if(I(t)===t&&!o&&!c){if(n=b.get(t),n)return n;if(n=++wt+"~",b.set(t,n),Array.isArray(t)){for(n="@",s=0;s<t.length;s++)n+=M(t[s])+",";b.set(t,n)}if(i){n="#";const a=I.keys(t).sort();for(;!p(s=a.pop());)p(t[s])||(n+=s+":"+M(t[s])+",");b.set(t,n)}}else n=o?t.toJSON():e=="symbol"?t.toString():e=="string"?JSON.stringify(t):""+t;return n},tt=t=>{if(h(t))try{t=t()}catch{t=""}const e=t;return t=typeof t=="string"?t:(Array.isArray(t)?t.length:t)?M(t):"",[t,e]};let _t=0;const Y=()=>++_t,vt=0,Ot=1,Ct=2;async function Rt(...t){const[e,o,c,i]=t,n=S({populateCache:!0,throwOnError:!0},typeof i=="boolean"?{revalidate:i}:i||{});let s=n.populateCache;const a=n.rollbackOnError;let r=n.optimisticData;const g=f=>typeof a=="function"?a(f):a!==!1,m=n.throwOnError;if(h(o)){const f=o,u=[],O=e.keys();for(const w of O)!/^\$(inf|sub)\$/.test(w)&&f(e.get(w)._k)&&u.push(w);return Promise.all(u.map(d))}return d(o);async function d(f){const[u]=tt(f);if(!u)return;const[O,w]=at(e,u),[ot,A,st,it]=R.get(e),J=()=>{const v=ot[u];return(h(n.revalidate)?n.revalidate(O().data,f):n.revalidate!==!1)&&(delete st[u],delete it[u],v&&v[0])?v[0](Ct).then(()=>O().data):O().data};if(t.length<3)return J();let E=c,_;const K=Y();A[u]=[K,0];const q=!p(r),B=O(),z=B.data,G=B._c,D=p(G)?z:G;if(q&&(r=h(r)?r(D,z):r,w({data:r,_c:D})),h(E))try{E=E(D)}catch(v){_=v}if(E&&rt(E))if(E=await E.catch(v=>{_=v}),K!==A[u][0]){if(_)throw _;return E}else _&&q&&g(_)&&(s=!0,w({data:D,_c:l}));if(s&&!_)if(h(s)){const v=s(E,D);w({data:v,error:l,_c:l})}else w({data:E,error:l,_c:l});if(A[u][1]=Y(),Promise.resolve(J()).then(()=>{w({_c:l})}),_){if(m)throw _;return}return E}}const Q=(t,e)=>{for(const o in t)t[o][0]&&t[o][0](e)},et=(t,e)=>{if(!R.has(t)){const o=S(pt,e),c={},i=Rt.bind(l,t);let n=T;const s={},a=(m,d)=>{const f=s[m]||[];return s[m]=f,f.push(d),()=>f.splice(f.indexOf(d),1)},r=(m,d,f)=>{t.set(m,d);const u=s[m];if(u)for(const O of u)O(d,f)},g=()=>{if(!R.has(t)&&(R.set(t,[c,{},{},{},i,r,a]),!$)){const m=o.initFocus(setTimeout.bind(l,Q.bind(l,c,vt))),d=o.initReconnect(setTimeout.bind(l,Q.bind(l,c,Ot)));n=()=>{m&&m(),d&&d(),R.delete(t)}}};return g(),[t,i,g,n]}return[t,R.get(t)[4]]},ht=(t,e,o,c,i)=>{const n=o.errorRetryCount,s=i.retryCount,a=~~((Math.random()+.5)*(1<<(s<8?s:8)))*o.errorRetryInterval;!p(n)&&s>n||setTimeout(c,a,i)},Tt=ct,[H,St]=et(new Map),Dt=S({onLoadingSlow:T,onSuccess:T,onError:T,onErrorRetry:ht,onDiscarded:T,revalidateOnFocus:!0,revalidateOnReconnect:!0,revalidateIfStale:!0,shouldRetryOnError:!0,errorRetryInterval:X?1e4:5e3,focusThrottleInterval:5*1e3,dedupingInterval:2*1e3,loadingTimeout:X?5e3:3e3,compare:Tt,isPaused:()=>!1,cache:H,mutate:St,fallback:{}},Et),gt=(t,e)=>{const o=S(t,e);if(e){const{use:c,fallback:i}=t,{use:n,fallback:s}=e;c&&n&&(o.use=c.concat(n)),i&&s&&(o.fallback=S(i,s))}return o},Z=C.createContext({}),yt=t=>{const{value:e}=t,o=C.useContext(Z),c=h(e),i=C.useMemo(()=>c?e(o):e,[c,o,e]),n=C.useMemo(()=>c?i:gt(o,i),[c,o,i]),s=i&&i.provider,a=C.useRef(l);s&&!a.current&&(a.current=et(s(n.cache||H),i));const r=a.current;return r&&(n.cache=r[0],n.mutate=r[1]),mt(()=>{if(r)return r[2]&&r[2](),r[3]},[]),C.createElement(Z.Provider,S(t,{value:n}))},bt="$inf$",nt=U&&window.__SWR_DEVTOOLS_USE__,It=nt?window.__SWR_DEVTOOLS_USE__:[],At=()=>{nt&&(window.__SWR_DEVTOOLS_REACT__=j)},Lt=t=>(e,o,c)=>t(e,o&&((...n)=>{const[s]=tt(e),[,,,a]=R.get(H);if(s.startsWith(bt))return o(...n);const r=a[s];return p(r)?o(...n):(delete a[s],r)}),c);It.concat(Lt);At();j.use;const Wt=I.defineProperty(yt,"defaultValue",{value:Dt});export{Wt as S};
