--- assets/vendor_system-DdpyrVMS.js
@@ -1,28 +1,0 @@
-import{r as y,a as ft,T as dt,j as B,R as It}from"./vendor_react-BO86OV0u.js";import{G as mt,s as ht,i as nt,a as Nt}from"./vendor_styled-engine-BqETdHfM.js";import{c as pt}from"./vendor_clsx-B-dksMZM.js";import{u as gt,T as Wt}from"./vendor_private-theming-CHDA6u9N.js";function je(e,...t){const n=new URL(`https://mui.com/production-error/?code=${e}`);return t.forEach(r=>n.searchParams.append("args[]",r)),`Minified MUI error #${e}; visit ${n} for the full message.`}function q(e){if(typeof e!="object"||e===null)return!1;const t=Object.getPrototypeOf(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)}function yt(e){if(y.isValidElement(e)||ft.isValidElementType(e)||!q(e))return e;const t={};return Object.keys(e).forEach(n=>{t[n]=yt(e[n])}),t}function Y(e,t,n={clone:!0}){const r=n.clone?{...e}:e;return q(e)&&q(t)&&Object.keys(t).forEach(o=>{y.isValidElement(t[o])||ft.isValidElementType(t[o])?r[o]=t[o]:q(t[o])&&Object.prototype.hasOwnProperty.call(e,o)&&q(e[o])?r[o]=Y(e[o],t[o],n):n.clone?r[o]=q(t[o])?yt(t[o]):t[o]:r[o]=t[o]}),r}const _t=e=>{const t=Object.keys(e).map(n=>({key:n,val:e[n]}))||[];return t.sort((n,r)=>n.val-r.val),t.reduce((n,r)=>({...n,[r.key]:r.val}),{})};function Gt(e){const{values:t={xs:0,sm:600,md:900,lg:1200,xl:1536},unit:n="px",step:r=5,...o}=e,s=_t(t),i=Object.keys(s);function c(u){return`@media (min-width:${typeof t[u]=="number"?t[u]:u}${n})`}function a(u){return`@media (max-width:${(typeof t[u]=="number"?t[u]:u)-r/100}${n})`}function l(u,g){const x=i.indexOf(g);return`@media (min-width:${typeof t[u]=="number"?t[u]:u}${n}) and (max-width:${(x!==-1&&typeof t[i[x]]=="number"?t[i[x]]:g)-r/100}${n})`}function f(u){return i.indexOf(u)+1<i.length?l(u,i[i.indexOf(u)+1]):c(u)}function m(u){const g=i.indexOf(u);return g===0?c(i[1]):g===i.length-1?a(i[g]):l(u,i[i.indexOf(u)+1]).replace("@media","@media not all and")}return{keys:i,values:s,up:c,down:a,between:l,only:f,not:m,unit:n,...o}}function Dt(e,t){if(!e.containerQueries)return t;const n=Object.keys(t).filter(r=>r.startsWith("@container")).sort((r,o)=>{var i,c;const s=/min-width:\s*([0-9.]+)/;return+(((i=r.match(s))==null?void 0:i[1])||0)-+(((c=o.match(s))==null?void 0:c[1])||0)});return n.length?n.reduce((r,o)=>{const s=t[o];return delete r[o],r[o]=s,r},{...t}):t}function Ft(e,t){return t==="@"||t.startsWith("@")&&(e.some(n=>t.startsWith(`@${n}`))||!!t.match(/^@\d/))}function Vt(e,t){const n=t.match(/^@([^/]+)?\/?(.+)?$/);if(!n)return null;const[,r,o]=n,s=Number.isNaN(+r)?r||0:+r;return e.containerQueries(o).up(s)}function zt(e){const t=(s,i)=>s.replace("@media",i?`@container ${i}`:"@container");function n(s,i){s.up=(...c)=>t(e.breakpoints.up(...c),i),s.down=(...c)=>t(e.breakpoints.down(...c),i),s.between=(...c)=>t(e.breakpoints.between(...c),i),s.only=(...c)=>t(e.breakpoints.only(...c),i),s.not=(...c)=>{const a=t(e.breakpoints.not(...c),i);return a.includes("not all and")?a.replace("not all and ","").replace("min-width:","width<").replace("max-width:","width>").replace("and","or"):a}}const r={},o=s=>(n(r,s),r);return n(o),{...e,containerQueries:o}}const Qt={borderRadius:4};function ce(e,t){return t?Y(e,t,{clone:!1}):e}const me={xs:0,sm:600,md:900,lg:1200,xl:1536},rt={keys:["xs","sm","md","lg","xl"],up:e=>`@media (min-width:${me[e]}px)`},Ut={containerQueries:e=>({up:t=>{let n=typeof t=="number"?t:me[t]||t;return typeof n=="number"&&(n=`${n}px`),e?`@container ${e} (min-width:${n})`:`@container (min-width:${n})`}})};function Q(e,t,n){const r=e.theme||{};if(Array.isArray(t)){const s=r.breakpoints||rt;return t.reduce((i,c,a)=>(i[s.up(s.keys[a])]=n(t[a]),i),{})}if(typeof t=="object"){const s=r.breakpoints||rt;return Object.keys(t).reduce((i,c)=>{if(Ft(s.keys,c)){const a=Vt(r.containerQueries?r:Ut,c);a&&(i[a]=n(t[c],c))}else if(Object.keys(s.values||me).includes(c)){const a=s.up(c);i[a]=n(t[c],c)}else{const a=c;i[a]=t[a]}return i},{})}return n(t)}function St(e={}){var n;return((n=e.keys)==null?void 0:n.reduce((r,o)=>{const s=e.up(o);return r[s]={},r},{}))||{}}function bt(e,t){return e.reduce((n,r)=>{const o=n[r];return(!o||Object.keys(o).length===0)&&delete n[r],n},t)}function Ht(e,...t){const n=St(e),r=[n,...t].reduce((o,s)=>Y(o,s),{});return bt(Object.keys(n),r)}function Yt(e,t){if(typeof e!="object")return{};const n={},r=Object.keys(t);return Array.isArray(e)?r.forEach((o,s)=>{s<e.length&&(n[o]=!0)}):r.forEach(o=>{e[o]!=null&&(n[o]=!0)}),n}function we({values:e,breakpoints:t,base:n}){const r=n||Yt(e,t),o=Object.keys(r);if(o.length===0)return e;let s;return o.reduce((i,c,a)=>(Array.isArray(e)?(i[c]=e[a]!=null?e[a]:e[s],s=a):typeof e=="object"?(i[c]=e[c]!=null?e[c]:e[s],s=c):i[c]=e,i),{})}function Ct(e){if(typeof e!="string")throw new Error(je(7));return e.charAt(0).toUpperCase()+e.slice(1)}function he(e,t,n=!0){if(!t||typeof t!="string")return null;if(e&&e.vars&&n){const r=`vars.${t}`.split(".").reduce((o,s)=>o&&o[s]?o[s]:null,e);if(r!=null)return r}return t.split(".").reduce((r,o)=>r&&r[o]!=null?r[o]:null,e)}function fe(e,t,n,r=n){let o;return typeof e=="function"?o=e(n):Array.isArray(e)?o=e[n]||r:o=he(e,n)||r,t&&(o=t(o,r,e)),o}function v(e){const{prop:t,cssProperty:n=e.prop,themeKey:r,transform:o}=e,s=i=>{if(i[t]==null)return null;const c=i[t],a=i.theme,l=he(a,r)||{};return Q(i,c,m=>{let u=fe(l,o,m);return m===u&&typeof m=="string"&&(u=fe(l,o,`${t}${m==="default"?"":Ct(m)}`,m)),n===!1?u:{[n]:u}})};return s.propTypes={},s.filterProps=[t],s}function Xt(e){const t={};return n=>(t[n]===void 0&&(t[n]=e(n)),t[n])}const qt={m:"margin",p:"padding"},Jt={t:"Top",r:"Right",b:"Bottom",l:"Left",x:["Left","Right"],y:["Top","Bottom"]},ot={marginX:"mx",marginY:"my",paddingX:"px",paddingY:"py"},Zt=Xt(e=>{if(e.length>2)if(ot[e])e=ot[e];else return[e];const[t,n]=e.split(""),r=qt[t],o=Jt[n]||"";return Array.isArray(o)?o.map(s=>r+s):[r+o]}),ke=["m","mt","mr","mb","ml","mx","my","margin","marginTop","marginRight","marginBottom","marginLeft","marginX","marginY","marginInline","marginInlineStart","marginInlineEnd","marginBlock","marginBlockStart","marginBlockEnd"],Oe=["p","pt","pr","pb","pl","px","py","padding","paddingTop","paddingRight","paddingBottom","paddingLeft","paddingX","paddingY","paddingInline","paddingInlineStart","paddingInlineEnd","paddingBlock","paddingBlockStart","paddingBlockEnd"];[...ke,...Oe];function le(e,t,n,r){const o=he(e,t,!0)??n;return typeof o=="number"||typeof o=="string"?s=>typeof s=="string"?s:typeof o=="string"?`calc(${s} * ${o})`:o*s:Array.isArray(o)?s=>{if(typeof s=="string")return s;const i=Math.abs(s),c=o[i];return s>=0?c:typeof c=="number"?-c:`-${c}`}:typeof o=="function"?o:()=>{}}function Ae(e){return le(e,"spacing",8)}function Z(e,t){return typeof t=="string"||t==null?t:e(t)}function en(e,t){return n=>e.reduce((r,o)=>(r[o]=Z(t,n),r),{})}function tn(e,t,n,r){if(!t.includes(n))return null;const o=Zt(n),s=en(o,r),i=e[n];return Q(e,i,s)}function xt(e,t){const n=Ae(e.theme);return Object.keys(e).map(r=>tn(e,t,r,n)).reduce(ce,{})}function $(e){return xt(e,ke)}$.propTypes={};$.filterProps=ke;function w(e){return xt(e,Oe)}w.propTypes={};w.filterProps=Oe;function nn(e=8,t=Ae({spacing:e})){if(e.mui)return e;const n=(...r)=>(r.length===0?[1]:r).map(s=>{const i=t(s);return typeof i=="number"?`${i}px`:i}).join(" ");return n.mui=!0,n}function pe(...e){const t=e.reduce((r,o)=>(o.filterProps.forEach(s=>{r[s]=o}),r),{}),n=r=>Object.keys(r).reduce((o,s)=>t[s]?ce(o,t[s](r)):o,{});return n.propTypes={},n.filterProps=e.reduce((r,o)=>r.concat(o.filterProps),[]),n}function G(e){return typeof e!="number"?e:`${e}px solid`}function D(e,t){return v({prop:e,themeKey:"borders",transform:t})}const rn=D("border",G),on=D("borderTop",G),sn=D("borderRight",G),cn=D("borderBottom",G),an=D("borderLeft",G),ln=D("borderColor"),un=D("borderTopColor"),fn=D("borderRightColor"),dn=D("borderBottomColor"),mn=D("borderLeftColor"),hn=D("outline",G),pn=D("outlineColor"),ge=e=>{if(e.borderRadius!==void 0&&e.borderRadius!==null){const t=le(e.theme,"shape.borderRadius",4),n=r=>({borderRadius:Z(t,r)});return Q(e,e.borderRadius,n)}return null};ge.propTypes={};ge.filterProps=["borderRadius"];pe(rn,on,sn,cn,an,ln,un,fn,dn,mn,ge,hn,pn);const ye=e=>{if(e.gap!==void 0&&e.gap!==null){const t=le(e.theme,"spacing",8),n=r=>({gap:Z(t,r)});return Q(e,e.gap,n)}return null};ye.propTypes={};ye.filterProps=["gap"];const Se=e=>{if(e.columnGap!==void 0&&e.columnGap!==null){const t=le(e.theme,"spacing",8),n=r=>({columnGap:Z(t,r)});return Q(e,e.columnGap,n)}return null};Se.propTypes={};Se.filterProps=["columnGap"];const be=e=>{if(e.rowGap!==void 0&&e.rowGap!==null){const t=le(e.theme,"spacing",8),n=r=>({rowGap:Z(t,r)});return Q(e,e.rowGap,n)}return null};be.propTypes={};be.filterProps=["rowGap"];const gn=v({prop:"gridColumn"}),yn=v({prop:"gridRow"}),Sn=v({prop:"gridAutoFlow"}),bn=v({prop:"gridAutoColumns"}),Cn=v({prop:"gridAutoRows"}),xn=v({prop:"gridTemplateColumns"}),$n=v({prop:"gridTemplateRows"}),wn=v({prop:"gridTemplateAreas"}),vn=v({prop:"gridArea"});pe(ye,Se,be,gn,yn,Sn,bn,Cn,xn,$n,wn,vn);function ne(e,t){return t==="grey"?t:e}const Tn=v({prop:"color",themeKey:"palette",transform:ne}),Pn=v({prop:"bgcolor",cssProperty:"backgroundColor",themeKey:"palette",transform:ne}),jn=v({prop:"backgroundColor",themeKey:"palette",transform:ne});pe(Tn,Pn,jn);function K(e){return e<=1&&e!==0?`${e*100}%`:e}const En=v({prop:"width",transform:K}),Me=e=>{if(e.maxWidth!==void 0&&e.maxWidth!==null){const t=n=>{var o,s,i,c,a;const r=((i=(s=(o=e.theme)==null?void 0:o.breakpoints)==null?void 0:s.values)==null?void 0:i[n])||me[n];return r?((a=(c=e.theme)==null?void 0:c.breakpoints)==null?void 0:a.unit)!=="px"?{maxWidth:`${r}${e.theme.breakpoints.unit}`}:{maxWidth:r}:{maxWidth:K(n)}};return Q(e,e.maxWidth,t)}return null};Me.filterProps=["maxWidth"];const kn=v({prop:"minWidth",transform:K}),On=v({prop:"height",transform:K}),An=v({prop:"maxHeight",transform:K}),Mn=v({prop:"minHeight",transform:K});v({prop:"size",cssProperty:"width",transform:K});v({prop:"size",cssProperty:"height",transform:K});const Rn=v({prop:"boxSizing"});pe(En,Me,kn,On,An,Mn,Rn);const Re={border:{themeKey:"borders",transform:G},borderTop:{themeKey:"borders",transform:G},borderRight:{themeKey:"borders",transform:G},borderBottom:{themeKey:"borders",transform:G},borderLeft:{themeKey:"borders",transform:G},borderColor:{themeKey:"palette"},borderTopColor:{themeKey:"palette"},borderRightColor:{themeKey:"palette"},borderBottomColor:{themeKey:"palette"},borderLeftColor:{themeKey:"palette"},outline:{themeKey:"borders",transform:G},outlineColor:{themeKey:"palette"},borderRadius:{themeKey:"shape.borderRadius",style:ge},color:{themeKey:"palette",transform:ne},bgcolor:{themeKey:"palette",cssProperty:"backgroundColor",transform:ne},backgroundColor:{themeKey:"palette",transform:ne},p:{style:w},pt:{style:w},pr:{style:w},pb:{style:w},pl:{style:w},px:{style:w},py:{style:w},padding:{style:w},paddingTop:{style:w},paddingRight:{style:w},paddingBottom:{style:w},paddingLeft:{style:w},paddingX:{style:w},paddingY:{style:w},paddingInline:{style:w},paddingInlineStart:{style:w},paddingInlineEnd:{style:w},paddingBlock:{style:w},paddingBlockStart:{style:w},paddingBlockEnd:{style:w},m:{style:$},mt:{style:$},mr:{style:$},mb:{style:$},ml:{style:$},mx:{style:$},my:{style:$},margin:{style:$},marginTop:{style:$},marginRight:{style:$},marginBottom:{style:$},marginLeft:{style:$},marginX:{style:$},marginY:{style:$},marginInline:{style:$},marginInlineStart:{style:$},marginInlineEnd:{style:$},marginBlock:{style:$},marginBlockStart:{style:$},marginBlockEnd:{style:$},displayPrint:{cssProperty:!1,transform:e=>({"@media print":{display:e}})},display:{},overflow:{},textOverflow:{},visibility:{},whiteSpace:{},flexBasis:{},flexDirection:{},flexWrap:{},justifyContent:{},alignItems:{},alignContent:{},order:{},flex:{},flexGrow:{},flexShrink:{},alignSelf:{},justifyItems:{},justifySelf:{},gap:{style:ye},rowGap:{style:be},columnGap:{style:Se},gridColumn:{},gridRow:{},gridAutoFlow:{},gridAutoColumns:{},gridAutoRows:{},gridTemplateColumns:{},gridTemplateRows:{},gridTemplateAreas:{},gridArea:{},position:{},zIndex:{themeKey:"zIndex"},top:{},right:{},bottom:{},left:{},boxShadow:{themeKey:"shadows"},width:{transform:K},maxWidth:{style:Me},minWidth:{transform:K},height:{transform:K},maxHeight:{transform:K},minHeight:{transform:K},boxSizing:{},font:{themeKey:"font"},fontFamily:{themeKey:"typography"},fontSize:{themeKey:"typography"},fontStyle:{themeKey:"typography"},fontWeight:{themeKey:"typography"},letterSpacing:{},textTransform:{},lineHeight:{},textAlign:{},typography:{cssProperty:!1,themeKey:"typography"}};function Bn(...e){const t=e.reduce((r,o)=>r.concat(Object.keys(o)),[]),n=new Set(t);return e.every(r=>n.size===Object.keys(r).length)}function Ln(e,t){return typeof e=="function"?e(t):e}function Kn(){function e(n,r,o,s){const i={[n]:r,theme:o},c=s[n];if(!c)return{[n]:r};const{cssProperty:a=n,themeKey:l,transform:f,style:m}=c;if(r==null)return null;if(l==="typography"&&r==="inherit")return{[n]:r};const u=he(o,l)||{};return m?m(i):Q(i,r,x=>{let T=fe(u,f,x);return x===T&&typeof x=="string"&&(T=fe(u,f,`${n}${x==="default"?"":Ct(x)}`,x)),a===!1?T:{[a]:T}})}function t(n){const{sx:r,theme:o={}}=n||{};if(!r)return null;const s=o.unstable_sxConfig??Re;function i(c){let a=c;if(typeof c=="function")a=c(o);else if(typeof c!="object")return c;if(!a)return null;const l=St(o.breakpoints),f=Object.keys(l);let m=l;return Object.keys(a).forEach(u=>{const g=Ln(a[u],o);if(g!=null)if(typeof g=="object")if(s[u])m=ce(m,e(u,g,o,s));else{const x=Q({theme:o},g,T=>({[u]:T}));Bn(x,g)?m[u]=t({sx:g,theme:o}):m=ce(m,x)}else m=ce(m,e(u,g,o,s))}),Dt(o,bt(f,m))}return Array.isArray(r)?r.map(i):i(r)}return t}const ae=Kn();ae.filterProps=["sx"];function In(e,t){var r;const n=this;if(n.vars){if(!((r=n.colorSchemes)!=null&&r[e])||typeof n.getColorSchemeSelector!="function")return{};let o=n.getColorSchemeSelector(e);return o==="&"?t:((o.includes("data-")||o.includes("."))&&(o=`*:where(${o.replace(/\s*&$/,"")}) &`),{[o]:t})}return n.palette.mode===e?t:{}}function Be(e={},...t){const{breakpoints:n={},palette:r={},spacing:o,shape:s={},...i}=e,c=Gt(n),a=nn(o);let l=Y({breakpoints:c,direction:"ltr",components:{},palette:{mode:"light",...r},spacing:a,shape:{...Qt,...s}},i);return l=zt(l),l.applyStyles=In,l=t.reduce((f,m)=>Y(f,m),l),l.unstable_sxConfig={...Re,...i==null?void 0:i.unstable_sxConfig},l.unstable_sx=function(m){return ae({sx:m,theme:this})},l}function Nn(e){return Object.keys(e).length===0}function Le(e=null){const t=y.useContext(dt);return!t||Nn(t)?e:t}const Wn=Be();function Ke(e=Wn){return Le(e)}function Mr({styles:e,themeId:t,defaultTheme:n={}}){const r=Ke(n),o=typeof e=="function"?e(t&&r[t]||r):e;return B.jsx(mt,{styles:o})}const _n=e=>{var r;const t={systemProps:{},otherProps:{}},n=((r=e==null?void 0:e.theme)==null?void 0:r.unstable_sxConfig)??Re;return Object.keys(e).forEach(o=>{n[o]?t.systemProps[o]=e[o]:t.otherProps[o]=e[o]}),t};function $t(e){const{sx:t,...n}=e,{systemProps:r,otherProps:o}=_n(n);let s;return Array.isArray(t)?s=[r,...t]:typeof t=="function"?s=(...i)=>{const c=t(...i);return q(c)?{...r,...c}:r}:s={...r,...t},{...o,sx:s}}const st=e=>e,Gn=()=>{let e=st;return{configure(t){e=t},generate(t){return e(t)},reset(){e=st}}},Dn=Gn();function Rr(e={}){const{themeId:t,defaultTheme:n,defaultClassName:r="MuiBox-root",generateClassName:o}=e,s=ht("div",{shouldForwardProp:c=>c!=="theme"&&c!=="sx"&&c!=="as"})(ae);return y.forwardRef(function(a,l){const f=Ke(n),{className:m,component:u="div",...g}=$t(a);return B.jsx(s,{as:u,ref:l,className:pt(m,o?o(r):r),theme:t&&f[t]||f,...g})})}const Fn={active:"active",checked:"checked",completed:"completed",disabled:"disabled",error:"error",expanded:"expanded",focused:"focused",focusVisible:"focusVisible",open:"open",readOnly:"readOnly",required:"required",selected:"selected"};function Vn(e,t,n="Mui"){const r=Fn[t];return r?`${n}-${r}`:`${Dn.generate(e)}-${t}`}function wt(e){const{variants:t,...n}=e,r={variants:t,style:nt(n),isProcessed:!0};return r.style===n||t&&t.forEach(o=>{typeof o.style!="function"&&(o.style=nt(o.style))}),r}const zn=Be();function ve(e){return e!=="ownerState"&&e!=="theme"&&e!=="sx"&&e!=="as"}function Qn(e){return e?(t,n)=>n[e]:null}function Un(e,t,n){e.theme=Xn(e.theme)?n:e.theme[t]||e.theme}function ue(e,t){const n=typeof t=="function"?t(e):t;if(Array.isArray(n))return n.flatMap(r=>ue(e,r));if(Array.isArray(n==null?void 0:n.variants)){let r;if(n.isProcessed)r=n.style;else{const{variants:o,...s}=n;r=s}return vt(e,n.variants,[r])}return n!=null&&n.isProcessed?n.style:n}function vt(e,t,n=[]){var o;let r;e:for(let s=0;s<t.length;s+=1){const i=t[s];if(typeof i.props=="function"){if(r??(r={...e,...e.ownerState,ownerState:e.ownerState}),!i.props(r))continue}else for(const c in i.props)if(e[c]!==i.props[c]&&((o=e.ownerState)==null?void 0:o[c])!==i.props[c])continue e;typeof i.style=="function"?(r??(r={...e,...e.ownerState,ownerState:e.ownerState}),n.push(i.style(r))):n.push(i.style)}return n}function Hn(e={}){const{themeId:t,defaultTheme:n=zn,rootShouldForwardProp:r=ve,slotShouldForwardProp:o=ve}=e;function s(c){Un(c,t,n)}return(c,a={})=>{Nt(c,d=>d.filter(h=>h!==ae));const{name:l,slot:f,skipVariantsResolver:m,skipSx:u,overridesResolver:g=Qn(Jn(f)),...x}=a,T=m!==void 0?m:f&&f!=="Root"&&f!=="root"||!1,F=u||!1;let E=ve;f==="Root"||f==="root"?E=r:f?E=o:qn(c)&&(E=void 0);const V=ht(c,{shouldForwardProp:E,label:Yn(),...x}),S=d=>{if(typeof d=="function"&&d.__emotion_real!==d)return function(b){return ue(b,d)};if(q(d)){const h=wt(d);return h.variants?function(C){return ue(C,h)}:h.style}return d},p=(...d)=>{const h=[],b=d.map(S),C=[];if(h.push(s),l&&g&&C.push(function(P){var U,ee;const j=(ee=(U=P.theme.components)==null?void 0:U[l])==null?void 0:ee.styleOverrides;if(!j)return null;const N={};for(const re in j)N[re]=ue(P,j[re]);return g(P,N)}),l&&!T&&C.push(function(P){var N,U;const I=P.theme,j=(U=(N=I==null?void 0:I.components)==null?void 0:N[l])==null?void 0:U.variants;return j?vt(P,j):null}),F||C.push(ae),Array.isArray(b[0])){const k=b.shift(),P=new Array(h.length).fill(""),I=new Array(C.length).fill("");let j;j=[...P,...k,...I],j.raw=[...P,...k.raw,...I],h.unshift(j)}const A=[...h,...b,...C],M=V(...A);return c.muiName&&(M.muiName=c.muiName),M};return V.withConfig&&(p.withConfig=V.withConfig),p}}function Yn(e,t){return void 0}function Xn(e){for(const t in e)return!1;return!0}function qn(e){return typeof e=="string"&&e.charCodeAt(0)>96}function Jn(e){return e&&e.charAt(0).toLowerCase()+e.slice(1)}const Zn=Hn();function de(e,t){const n={...t};for(const r in e)if(Object.prototype.hasOwnProperty.call(e,r)){const o=r;if(o==="components"||o==="slots")n[o]={...e[o],...n[o]};else if(o==="componentsProps"||o==="slotProps"){const s=e[o],i=t[o];if(!i)n[o]=s||{};else if(!s)n[o]=i;else{n[o]={...i};for(const c in s)if(Object.prototype.hasOwnProperty.call(s,c)){const a=c;n[o][a]=de(s[a],i[a])}}}else n[o]===void 0&&(n[o]=e[o])}return n}function Tt(e){const{theme:t,name:n,props:r}=e;return!t||!t.components||!t.components[n]||!t.components[n].defaultProps?r:de(t.components[n].defaultProps,r)}function er({props:e,name:t,defaultTheme:n,themeId:r}){let o=Ke(n);return r&&(o=o[r]||o),Tt({theme:o,name:t,props:e})}const Pt=typeof window<"u"?y.useLayoutEffect:y.useEffect;function tr(e,t,n,r,o){const[s,i]=y.useState(()=>o&&n?n(e).matches:r?r(e).matches:t);return Pt(()=>{if(!n)return;const c=n(e),a=()=>{i(c.matches)};return a(),c.addEventListener("change",a),()=>{c.removeEventListener("change",a)}},[e,n]),s}const nr={...It},jt=nr.useSyncExternalStore;function rr(e,t,n,r,o){const s=y.useCallback(()=>t,[t]),i=y.useMemo(()=>{if(o&&n)return()=>n(e).matches;if(r!==null){const{matches:f}=r(e);return()=>f}return s},[s,e,r,o,n]),[c,a]=y.useMemo(()=>{if(n===null)return[s,()=>()=>{}];const f=n(e);return[()=>f.matches,m=>(f.addEventListener("change",m),()=>{f.removeEventListener("change",m)})]},[s,n,e]);return jt(a,c,i)}function Br(e={}){const{themeId:t}=e;return function(r,o={}){let s=Le();s&&t&&(s=s[t]||s);const i=typeof window<"u"&&typeof window.matchMedia<"u",{defaultMatches:c=!1,matchMedia:a=i?window.matchMedia:null,ssrMatchMedia:l=null,noSsr:f=!1}=Tt({name:"MuiUseMediaQuery",props:o,theme:s});let m=typeof r=="function"?r(s):r;return m=m.replace(/^@media( ?)/m,""),(jt!==void 0?rr:tr)(m,c,a,l,f)}}function or(e,t=Number.MIN_SAFE_INTEGER,n=Number.MAX_SAFE_INTEGER){return Math.max(t,Math.min(e,n))}function Ie(e,t=0,n=1){return or(e,t,n)}function sr(e){e=e.slice(1);const t=new RegExp(`.{1,${e.length>=6?2:1}}`,"g");let n=e.match(t);return n&&n[0].length===1&&(n=n.map(r=>r+r)),n?`rgb${n.length===4?"a":""}(${n.map((r,o)=>o<3?parseInt(r,16):Math.round(parseInt(r,16)/255*1e3)/1e3).join(", ")})`:""}function J(e){if(e.type)return e;if(e.charAt(0)==="#")return J(sr(e));const t=e.indexOf("("),n=e.substring(0,t);if(!["rgb","rgba","hsl","hsla","color"].includes(n))throw new Error(je(9,e));let r=e.substring(t+1,e.length-1),o;if(n==="color"){if(r=r.split(" "),o=r.shift(),r.length===4&&r[3].charAt(0)==="/"&&(r[3]=r[3].slice(1)),!["srgb","display-p3","a98-rgb","prophoto-rgb","rec-2020"].includes(o))throw new Error(je(10,o))}else r=r.split(",");return r=r.map(s=>parseFloat(s)),{type:n,values:r,colorSpace:o}}const ir=e=>{const t=J(e);return t.values.slice(0,3).map((n,r)=>t.type.includes("hsl")&&r!==0?`${n}%`:n).join(" ")},Lr=(e,t)=>{try{return ir(e)}catch{return e}};function Ce(e){const{type:t,colorSpace:n}=e;let{values:r}=e;return t.includes("rgb")?r=r.map((o,s)=>s<3?parseInt(o,10):o):t.includes("hsl")&&(r[1]=`${r[1]}%`,r[2]=`${r[2]}%`),t.includes("color")?r=`${n} ${r.join(" ")}`:r=`${r.join(", ")}`,`${t}(${r})`}function cr(e){e=J(e);const{values:t}=e,n=t[0],r=t[1]/100,o=t[2]/100,s=r*Math.min(o,1-o),i=(l,f=(l+n/30)%12)=>o-s*Math.max(Math.min(f-3,9-f,1),-1);let c="rgb";const a=[Math.round(i(0)*255),Math.round(i(8)*255),Math.round(i(4)*255)];return e.type==="hsla"&&(c+="a",a.push(t[3])),Ce({type:c,values:a})}function Ee(e){e=J(e);let t=e.type==="hsl"||e.type==="hsla"?J(cr(e)).values:e.values;return t=t.map(n=>(e.type!=="color"&&(n/=255),n<=.03928?n/12.92:((n+.055)/1.055)**2.4)),Number((.2126*t[0]+.7152*t[1]+.0722*t[2]).toFixed(3))}function Kr(e,t){const n=Ee(e),r=Ee(t);return(Math.max(n,r)+.05)/(Math.min(n,r)+.05)}function ar(e,t){return e=J(e),t=Ie(t),(e.type==="rgb"||e.type==="hsl")&&(e.type+="a"),e.type==="color"?e.values[3]=`/${t}`:e.values[3]=t,Ce(e)}function Ir(e,t,n){try{return ar(e,t)}catch{return e}}function Et(e,t){if(e=J(e),t=Ie(t),e.type.includes("hsl"))e.values[2]*=1-t;else if(e.type.includes("rgb")||e.type.includes("color"))for(let n=0;n<3;n+=1)e.values[n]*=1-t;return Ce(e)}function Nr(e,t,n){try{return Et(e,t)}catch{return e}}function kt(e,t){if(e=J(e),t=Ie(t),e.type.includes("hsl"))e.values[2]+=(100-e.values[2])*t;else if(e.type.includes("rgb"))for(let n=0;n<3;n+=1)e.values[n]+=(255-e.values[n])*t;else if(e.type.includes("color"))for(let n=0;n<3;n+=1)e.values[n]+=(1-e.values[n])*t;return Ce(e)}function Wr(e,t,n){try{return kt(e,t)}catch{return e}}function lr(e,t=.15){return Ee(e)>.5?Et(e,t):kt(e,t)}function _r(e,t,n){try{return lr(e,t)}catch{return e}}const Ot=y.createContext();function ur({value:e,...t}){return B.jsx(Ot.Provider,{value:e??!0,...t})}const Gr=()=>y.useContext(Ot)??!1,At=y.createContext(void 0);function fr({value:e,children:t}){return B.jsx(At.Provider,{value:e,children:t})}function dr(e){const{theme:t,name:n,props:r}=e;if(!t||!t.components||!t.components[n])return r;const o=t.components[n];return o.defaultProps?de(o.defaultProps,r):!o.styleOverrides&&!o.variants?de(o,r):r}function Dr({props:e,name:t}){const n=y.useContext(At);return dr({props:e,name:t,theme:{components:n}})}const it={};function ct(e,t,n,r=!1){return y.useMemo(()=>{const o=e&&t[e]||t;if(typeof n=="function"){const s=n(o),i=e?{...t,[e]:s}:s;return r?()=>i:i}return e?{...t,[e]:n}:{...t,...n}},[e,t,n,r])}function mr(e){const{children:t,theme:n,themeId:r}=e,o=Le(it),s=gt()||it,i=ct(r,o,n),c=ct(r,s,n,!0),a=(r?i[r]:i).direction==="rtl";return B.jsx(Wt,{theme:c,children:B.jsx(dt.Provider,{value:i,children:B.jsx(ur,{value:a,children:B.jsx(fr,{value:r?i[r].components:i.components,children:t})})})})}const at={theme:void 0};function Fr(e){let t,n;return function(o){let s=t;return(s===void 0||o.theme!==n)&&(at.theme=o.theme,s=wt(e(at)),t=s,n=o.theme),s}}const Ne="mode",We="color-scheme",hr="data-color-scheme";function pr(e){const{defaultMode:t="system",defaultLightColorScheme:n="light",defaultDarkColorScheme:r="dark",modeStorageKey:o=Ne,colorSchemeStorageKey:s=We,attribute:i=hr,colorSchemeNode:c="document.documentElement",nonce:a}=e;let l="",f=i;if(i==="class"&&(f=".%s"),i==="data"&&(f="[data-%s]"),f.startsWith(".")){const u=f.substring(1);l+=`${c}.classList.remove('${u}'.replace('%s', light), '${u}'.replace('%s', dark));
-      ${c}.classList.add('${u}'.replace('%s', colorScheme));`}const m=f.match(/\[([^\]]+)\]/);if(m){const[u,g]=m[1].split("=");g||(l+=`${c}.removeAttribute('${u}'.replace('%s', light));
-      ${c}.removeAttribute('${u}'.replace('%s', dark));`),l+=`
-      ${c}.setAttribute('${u}'.replace('%s', colorScheme), ${g?`${g}.replace('%s', colorScheme)`:'""'});`}else l+=`${c}.setAttribute('${f}', colorScheme);`;return B.jsx("script",{suppressHydrationWarning:!0,nonce:typeof window>"u"?a:"",dangerouslySetInnerHTML:{__html:`(function() {
-try {
-  let colorScheme = '';
-  const mode = localStorage.getItem('${o}') || '${t}';
-  const dark = localStorage.getItem('${s}-dark') || '${r}';
-  const light = localStorage.getItem('${s}-light') || '${n}';
-  if (mode === 'system') {
-    // handle system mode
-    const mql = window.matchMedia('(prefers-color-scheme: dark)');
-    if (mql.matches) {
-      colorScheme = dark
-    } else {
-      colorScheme = light
-    }
-  }
-  if (mode === 'light') {
-    colorScheme = light;
-  }
-  if (mode === 'dark') {
-    colorScheme = dark;
-  }
-  if (colorScheme) {
-    ${l}
-  }
-} catch(e){}})();`}},"mui-color-scheme-init")}function lt(e){if(typeof window<"u"&&typeof window.matchMedia=="function"&&e==="system")return window.matchMedia("(prefers-color-scheme: dark)").matches?"dark":"light"}function Mt(e,t){if(e.mode==="light"||e.mode==="system"&&e.systemMode==="light")return t("light");if(e.mode==="dark"||e.mode==="system"&&e.systemMode==="dark")return t("dark")}function gr(e){return Mt(e,t=>{if(t==="light")return e.lightColorScheme;if(t==="dark")return e.darkColorScheme})}function Te(e,t){if(typeof window>"u")return;let n;try{n=localStorage.getItem(e)||void 0,n||localStorage.setItem(e,t)}catch{}return n||t}function yr(e){const{defaultMode:t="light",defaultLightColorScheme:n,defaultDarkColorScheme:r,supportedColorSchemes:o=[],modeStorageKey:s=Ne,colorSchemeStorageKey:i=We,storageWindow:c=typeof window>"u"?void 0:window,noSsr:a=!1}=e,l=o.join(","),f=o.length>1,[m,u]=y.useState(()=>{const p=Te(s,t),d=Te(`${i}-light`,n),h=Te(`${i}-dark`,r);return{mode:p,systemMode:lt(p),lightColorScheme:d,darkColorScheme:h}}),[g,x]=y.useState(a||!f);y.useEffect(()=>{x(!0)},[]);const T=gr(m),F=y.useCallback(p=>{u(d=>{if(p===d.mode)return d;const h=p??t;try{localStorage.setItem(s,h)}catch{}return{...d,mode:h,systemMode:lt(h)}})},[s,t]),E=y.useCallback(p=>{p?typeof p=="string"?p&&!l.includes(p)?console.error(`\`${p}\` does not exist in \`theme.colorSchemes\`.`):u(d=>{const h={...d};return Mt(d,b=>{try{localStorage.setItem(`${i}-${b}`,p)}catch{}b==="light"&&(h.lightColorScheme=p),b==="dark"&&(h.darkColorScheme=p)}),h}):u(d=>{const h={...d},b=p.light===null?n:p.light,C=p.dark===null?r:p.dark;if(b)if(!l.includes(b))console.error(`\`${b}\` does not exist in \`theme.colorSchemes\`.`);else{h.lightColorScheme=b;try{localStorage.setItem(`${i}-light`,b)}catch{}}if(C)if(!l.includes(C))console.error(`\`${C}\` does not exist in \`theme.colorSchemes\`.`);else{h.darkColorScheme=C;try{localStorage.setItem(`${i}-dark`,C)}catch{}}return h}):u(d=>{try{localStorage.setItem(`${i}-light`,n),localStorage.setItem(`${i}-dark`,r)}catch{}return{...d,lightColorScheme:n,darkColorScheme:r}})},[l,i,n,r]),V=y.useCallback(p=>{m.mode==="system"&&u(d=>{const h=p!=null&&p.matches?"dark":"light";return d.systemMode===h?d:{...d,systemMode:h}})},[m.mode]),S=y.useRef(V);return S.current=V,y.useEffect(()=>{if(typeof window.matchMedia!="function"||!f)return;const p=(...h)=>S.current(...h),d=window.matchMedia("(prefers-color-scheme: dark)");return d.addListener(p),p(d),()=>{d.removeListener(p)}},[f]),y.useEffect(()=>{if(c&&f){const p=d=>{const h=d.newValue;typeof d.key=="string"&&d.key.startsWith(i)&&(!h||l.match(h))&&(d.key.endsWith("light")&&E({light:h}),d.key.endsWith("dark")&&E({dark:h})),d.key===s&&(!h||["light","dark","system"].includes(h))&&F(h||t)};return c.addEventListener("storage",p),()=>{c.removeEventListener("storage",p)}}},[E,F,s,i,l,t,c,f]),{...m,mode:g?m.mode:void 0,systemMode:g?m.systemMode:void 0,colorScheme:g?T:void 0,setMode:F,setColorScheme:E}}const Sr="*{-webkit-transition:none!important;-moz-transition:none!important;-o-transition:none!important;-ms-transition:none!important;transition:none!important}";function Vr(e){const{themeId:t,theme:n={},modeStorageKey:r=Ne,colorSchemeStorageKey:o=We,disableTransitionOnChange:s=!1,defaultColorScheme:i,resolveTheme:c}=e,a={allColorSchemes:[],colorScheme:void 0,darkColorScheme:void 0,lightColorScheme:void 0,mode:void 0,setColorScheme:()=>{},setMode:()=>{},systemMode:void 0},l=y.createContext(void 0),f=()=>y.useContext(l)||a,m={},u={};function g(E){var Je,Ze,et,tt;const{children:V,theme:S,modeStorageKey:p=r,colorSchemeStorageKey:d=o,disableTransitionOnChange:h=s,storageWindow:b=typeof window>"u"?void 0:window,documentNode:C=typeof document>"u"?void 0:document,colorSchemeNode:A=typeof document>"u"?void 0:document.documentElement,disableNestedContext:M=!1,disableStyleSheetGeneration:k=!1,defaultMode:P="system",noSsr:I}=E,j=y.useRef(!1),N=gt(),U=y.useContext(l),ee=!!U&&!M,re=y.useMemo(()=>S||(typeof n=="function"?n():n),[S]),_e=re[t],W=_e||re,{colorSchemes:X=m,components:Ge=u,cssVarPrefix:xe}=W,De=Object.keys(X).filter(O=>!!X[O]).join(","),te=y.useMemo(()=>De.split(","),[De]),Fe=typeof i=="string"?i:i.light,Ve=typeof i=="string"?i:i.dark,Rt=X[Fe]&&X[Ve]?P:((Ze=(Je=X[W.defaultColorScheme])==null?void 0:Je.palette)==null?void 0:Ze.mode)||((et=W.palette)==null?void 0:et.mode),{mode:Bt,setMode:ze,systemMode:Qe,lightColorScheme:Ue,darkColorScheme:He,colorScheme:Lt,setColorScheme:Ye}=yr({supportedColorSchemes:te,defaultLightColorScheme:Fe,defaultDarkColorScheme:Ve,modeStorageKey:p,colorSchemeStorageKey:d,defaultMode:Rt,storageWindow:b,noSsr:I});let $e=Bt,_=Lt;ee&&($e=U.mode,_=U.colorScheme);const oe=y.useMemo(()=>{var ie;const O=_||W.defaultColorScheme,R=((ie=W.generateThemeVars)==null?void 0:ie.call(W))||W.vars,L={...W,components:Ge,colorSchemes:X,cssVarPrefix:xe,vars:R};if(typeof L.generateSpacing=="function"&&(L.spacing=L.generateSpacing()),O){const z=X[O];z&&typeof z=="object"&&Object.keys(z).forEach(H=>{z[H]&&typeof z[H]=="object"?L[H]={...L[H],...z[H]}:L[H]=z[H]})}return c?c(L):L},[W,_,Ge,X,xe]),se=W.colorSchemeSelector;Pt(()=>{if(_&&A&&se&&se!=="media"){const O=se;let R=se;if(O==="class"&&(R=".%s"),O==="data"&&(R="[data-%s]"),O!=null&&O.startsWith("data-")&&!O.includes("%s")&&(R=`[${O}="%s"]`),R.startsWith("."))A.classList.remove(...te.map(L=>R.substring(1).replace("%s",L))),A.classList.add(R.substring(1).replace("%s",_));else{const L=R.replace("%s",_).match(/\[([^\]]+)\]/);if(L){const[ie,z]=L[1].split("=");z||te.forEach(H=>{A.removeAttribute(ie.replace(_,H))}),A.setAttribute(ie,z?z.replace(/"|'/g,""):"")}else A.setAttribute(R,_)}}},[_,se,A,te]),y.useEffect(()=>{let O;if(h&&j.current&&C){const R=C.createElement("style");R.appendChild(C.createTextNode(Sr)),C.head.appendChild(R),window.getComputedStyle(C.body),O=setTimeout(()=>{C.head.removeChild(R)},1)}return()=>{clearTimeout(O)}},[_,h,C]),y.useEffect(()=>(j.current=!0,()=>{j.current=!1}),[]);const Kt=y.useMemo(()=>({allColorSchemes:te,colorScheme:_,darkColorScheme:He,lightColorScheme:Ue,mode:$e,setColorScheme:Ye,setMode:ze,systemMode:Qe}),[te,_,He,Ue,$e,Ye,ze,Qe,oe.colorSchemeSelector]);let Xe=!0;(k||W.cssVariables===!1||ee&&(N==null?void 0:N.cssVarPrefix)===xe)&&(Xe=!1);const qe=B.jsxs(y.Fragment,{children:[B.jsx(mr,{themeId:_e?t:void 0,theme:oe,children:V}),Xe&&B.jsx(mt,{styles:((tt=oe.generateStyleSheets)==null?void 0:tt.call(oe))||[]})]});return ee?qe:B.jsx(l.Provider,{value:Kt,children:qe})}const x=typeof i=="string"?i:i.light,T=typeof i=="string"?i:i.dark;return{CssVarsProvider:g,useColorScheme:f,getInitColorSchemeScript:E=>pr({colorSchemeStorageKey:o,defaultLightColorScheme:x,defaultDarkColorScheme:T,modeStorageKey:r,...E})}}function zr(e=""){function t(...r){if(!r.length)return"";const o=r[0];return typeof o=="string"&&!o.match(/(#|\(|\)|(-?(\d*\.)?\d+)(px|em|%|ex|ch|rem|vw|vh|vmin|vmax|cm|mm|in|pt|pc))|^(-?(\d*\.)?\d+)$|(\d+ \d+ \d+)/)?`, var(--${e?`${e}-`:""}${o}${t(...r.slice(1))})`:`, ${o}`}return(r,...o)=>`var(--${e?`${e}-`:""}${r}${t(...o)})`}const ut=(e,t,n,r=[])=>{let o=e;t.forEach((s,i)=>{i===t.length-1?Array.isArray(o)?o[Number(s)]=n:o&&typeof o=="object"&&(o[s]=n):o&&typeof o=="object"&&(o[s]||(o[s]=r.includes(s)?[]:{}),o=o[s])})},br=(e,t,n)=>{function r(o,s=[],i=[]){Object.entries(o).forEach(([c,a])=>{(!n||!n([...s,c]))&&a!=null&&(typeof a=="object"&&Object.keys(a).length>0?r(a,[...s,c],Array.isArray(a)?[...i,c]:i):t([...s,c],a,i))})}r(e)},Cr=(e,t)=>typeof t=="number"?["lineHeight","fontWeight","opacity","zIndex"].some(r=>e.includes(r))||e[e.length-1].toLowerCase().includes("opacity")?t:`${t}px`:t;function Pe(e,t){const{prefix:n,shouldSkipGeneratingVar:r}=t||{},o={},s={},i={};return br(e,(c,a,l)=>{if((typeof a=="string"||typeof a=="number")&&(!r||!r(c,a))){const f=`--${n?`${n}-`:""}${c.join("-")}`,m=Cr(c,a);Object.assign(o,{[f]:m}),ut(s,c,`var(${f})`,l),ut(i,c,`var(${f}, ${m})`,l)}},c=>c[0]==="vars"),{css:o,vars:s,varsWithDefaults:i}}function Qr(e,t={}){const{getSelector:n=F,disableCssColorScheme:r,colorSchemeSelector:o}=t,{colorSchemes:s={},components:i,defaultColorScheme:c="light",...a}=e,{vars:l,css:f,varsWithDefaults:m}=Pe(a,t);let u=m;const g={},{[c]:x,...T}=s;if(Object.entries(T||{}).forEach(([S,p])=>{const{vars:d,css:h,varsWithDefaults:b}=Pe(p,t);u=Y(u,b),g[S]={css:h,vars:d}}),x){const{css:S,vars:p,varsWithDefaults:d}=Pe(x,t);u=Y(u,d),g[c]={css:S,vars:p}}function F(S,p){var h,b;let d=o;if(o==="class"&&(d=".%s"),o==="data"&&(d="[data-%s]"),o!=null&&o.startsWith("data-")&&!o.includes("%s")&&(d=`[${o}="%s"]`),S){if(d==="media")return e.defaultColorScheme===S?":root":{[`@media (prefers-color-scheme: ${((b=(h=s[S])==null?void 0:h.palette)==null?void 0:b.mode)||S})`]:{":root":p}};if(d)return e.defaultColorScheme===S?`:root, ${d.replace("%s",String(S))}`:d.replace("%s",String(S))}return":root"}return{vars:u,generateThemeVars:()=>{let S={...l};return Object.entries(g).forEach(([,{vars:p}])=>{S=Y(S,p)}),S},generateStyleSheets:()=>{var C,A;const S=[],p=e.defaultColorScheme||"light";function d(M,k){Object.keys(k).length&&S.push(typeof M=="string"?{[M]:{...k}}:M)}d(n(void 0,{...f}),f);const{[p]:h,...b}=g;if(h){const{css:M}=h,k=(A=(C=s[p])==null?void 0:C.palette)==null?void 0:A.mode,P=!r&&k?{colorScheme:k,...M}:{...M};d(n(p,{...P}),P)}return Object.entries(b).forEach(([M,{css:k}])=>{var j,N;const P=(N=(j=s[M])==null?void 0:j.palette)==null?void 0:N.mode,I=!r&&P?{colorScheme:P,...k}:{...k};d(n(M,{...I}),I)}),S}}}function Ur(e){return function(n){return e==="media"?`@media (prefers-color-scheme: ${n})`:e?e.startsWith("data-")&&!e.includes("%s")?`[${e}="${n}"] &`:e==="class"?`.${n} &`:e==="data"?`[data-${n}] &`:`${e.replace("%s",n)} &`:"&"}}function xr(e,t,n=void 0){const r={};for(const o in e){const s=e[o];let i="",c=!0;for(let a=0;a<s.length;a+=1){const l=s[a];l&&(i+=(c===!0?"":" ")+t(l),c=!1,n&&n[l]&&(i+=" "+n[l]))}r[o]=i}return r}const $r=Be(),wr=Zn("div",{name:"MuiStack",slot:"Root",overridesResolver:(e,t)=>t.root});function vr(e){return er({props:e,name:"MuiStack",defaultTheme:$r})}function Tr(e,t){const n=y.Children.toArray(e).filter(Boolean);return n.reduce((r,o,s)=>(r.push(o),s<n.length-1&&r.push(y.cloneElement(t,{key:`separator-${s}`})),r),[])}const Pr=e=>({row:"Left","row-reverse":"Right",column:"Top","column-reverse":"Bottom"})[e],jr=({ownerState:e,theme:t})=>{let n={display:"flex",flexDirection:"column",...Q({theme:t},we({values:e.direction,breakpoints:t.breakpoints.values}),r=>({flexDirection:r}))};if(e.spacing){const r=Ae(t),o=Object.keys(t.breakpoints.values).reduce((a,l)=>((typeof e.spacing=="object"&&e.spacing[l]!=null||typeof e.direction=="object"&&e.direction[l]!=null)&&(a[l]=!0),a),{}),s=we({values:e.direction,base:o}),i=we({values:e.spacing,base:o});typeof s=="object"&&Object.keys(s).forEach((a,l,f)=>{if(!s[a]){const u=l>0?s[f[l-1]]:"column";s[a]=u}}),n=Y(n,Q({theme:t},i,(a,l)=>e.useFlexGap?{gap:Z(r,a)}:{"& > :not(style):not(style)":{margin:0},"& > :not(style) ~ :not(style)":{[`margin${Pr(l?s[l]:e.direction)}`]:Z(r,a)}}))}return n=Ht(t.breakpoints,n),n};function Hr(e={}){const{createStyledComponent:t=wr,useThemeProps:n=vr,componentName:r="MuiStack"}=e,o=()=>xr({root:["root"]},a=>Vn(r,a),{}),s=t(jr);return y.forwardRef(function(a,l){const f=n(a),m=$t(f),{component:u="div",direction:g="column",spacing:x=0,divider:T,children:F,className:E,useFlexGap:V=!1,...S}=m,p={direction:g,spacing:x,useFlexGap:V},d=o();return B.jsx(s,{as:u,ownerState:p,ref:l,className:pt(d.root,E),...S,children:T?Tr(F,T):F})})}function Yr(e){const t={};return Object.entries(e).forEach(r=>{const[o,s]=r;typeof s=="object"&&(t[o]=`${s.fontStyle?`${s.fontStyle} `:""}${s.fontVariant?`${s.fontVariant} `:""}${s.fontWeight?`${s.fontWeight} `:""}${s.fontStretch?`${s.fontStretch} `:""}${s.fontSize||""}${s.lineHeight?`/${s.lineHeight} `:""}${s.fontFamily||""}`)}),t}export{Hr as A,Rr as B,Br as C,Gr as D,Mr as G,mr as T,Re as a,Nr as b,Be as c,Et as d,Wr as e,_r as f,Kr as g,Ir as h,Lr as i,Qr as j,nn as k,kt as l,Ae as m,Ur as n,cr as o,Yr as p,zr as q,er as r,ae as s,Hn as t,Ke as u,Vr as v,$t as w,Fr as x,Dr as y,ar as z};
