--- assets/vendor_utils-DGLdK2NL.js
@@ -1,1 +1,0 @@
-import{a as c}from"./vendor_@babel-Brk9hZUz.js";import{r as i}from"./vendor_react-BO86OV0u.js";function f(e){let t="https://mui.com/production-error/?code="+e;for(let n=1;n<arguments.length;n+=1)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified MUI error #"+e+"; visit "+t+" for the full message."}function l(e){if(typeof e!="object"||e===null)return!1;const t=Object.getPrototypeOf(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)}function s(e){if(i.isValidElement(e)||!l(e))return e;const t={};return Object.keys(e).forEach(n=>{t[n]=s(e[n])}),t}function d(e,t,n={clone:!0}){const o=n.clone?c({},e):e;return l(e)&&l(t)&&Object.keys(t).forEach(r=>{i.isValidElement(t[r])?o[r]=t[r]:l(t[r])&&Object.prototype.hasOwnProperty.call(e,r)&&l(e[r])?o[r]=d(e[r],t[r],n):n.clone?o[r]=l(t[r])?s(t[r]):t[r]:o[r]=t[r]}),o}function O(e){if(typeof e!="string")throw new Error(f(7));return e.charAt(0).toUpperCase()+e.slice(1)}const a=e=>e,p=()=>{let e=a;return{configure(t){e=t},generate(t){return e(t)},reset(){e=a}}},u=p(),b={active:"active",checked:"checked",completed:"completed",disabled:"disabled",error:"error",expanded:"expanded",focused:"focused",focusVisible:"focusVisible",open:"open",readOnly:"readOnly",required:"required",selected:"selected"};function h(e,t,n="Mui"){const o=b[t];return o?`${n}-${o}`:`${u.generate(e)}-${t}`}export{O as c,d,f,h as g};
