+++ assets/vendor_x-date-pickers-BLC6f5Ep.js
@@ -1,4 +1,4 @@
-import{_ as O,a as m}from"./vendor_@babel-Brk9hZUz.js";import{r as c,j as P}from"./vendor_react-BO86OV0u.js";import{u as j}from"./vendor_material-D1sBINAn.js";import{d as n,l as w,w as S,i as C,a as Y,c as z}from"./vendor_dayjs-DVEZ5JDk.js";const A=["localeText"],b=c.createContext(null),B=function(r){const{localeText:o}=r,e=O(r,A),{utils:t,localeText:i}=c.useContext(b)??{utils:void 0,localeText:void 0},d=j({props:e,name:"MuiLocalizationProvider"}),{children:h,dateAdapter:a,dateFormats:l,dateLibInstance:T,adapterLocale:y,localeText:M}=d,p=c.useMemo(()=>m({},M,i,o),[M,i,o]),u=c.useMemo(()=>{if(!a)return t||null;const D=new a({locale:y,formats:l,instance:T});if(!D.isMUIAdapter)throw new Error(["MUI X: The date adapter should be imported from `@mui/x-date-pickers` or `@mui/x-date-pickers-pro`, not from `@date-io`","For example, `import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'` instead of `import AdapterDayjs from '@date-io/dayjs'`","More information on the installation documentation: https://mui.com/x/react-date-pickers/getting-started/#installation"].join(`
+import{_ as O,a as m}from"./vendor_@babel-BRdSvYXJ.js";import{r as c,j as P}from"./vendor_react-BJ2hNBYs.js";import{u as j}from"./vendor_material-GI4Anplu.js";import{d as n,l as w,w as S,i as C,a as Y,c as z}from"./vendor_dayjs-DVEZ5JDk.js";const A=["localeText"],b=c.createContext(null),B=function(r){const{localeText:o}=r,e=O(r,A),{utils:t,localeText:i}=c.useContext(b)??{utils:void 0,localeText:void 0},d=j({props:e,name:"MuiLocalizationProvider"}),{children:h,dateAdapter:a,dateFormats:l,dateLibInstance:T,adapterLocale:y,localeText:M}=d,p=c.useMemo(()=>m({},M,i,o),[M,i,o]),u=c.useMemo(()=>{if(!a)return t||null;const D=new a({locale:y,formats:l,instance:T});if(!D.isMUIAdapter)throw new Error(["MUI X: The date adapter should be imported from `@mui/x-date-pickers` or `@mui/x-date-pickers-pro`, not from `@date-io`","For example, `import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'` instead of `import AdapterDayjs from '@date-io/dayjs'`","More information on the installation documentation: https://mui.com/x/react-date-pickers/getting-started/#installation"].join(`
 `));return D},[a,y,l,T,t]),g=c.useMemo(()=>u?{minDate:u.date("1900-01-01T00:00:00.000"),maxDate:u.date("2099-12-31T00:00:00.000")}:null,[u]),x=c.useMemo(()=>({utils:u,defaultDates:g,localeText:p}),[g,u,p]);return P.jsx(b.Provider,{value:x,children:h})};n.extend(w);n.extend(S);n.extend(C);n.extend(Y);const U={YY:"year",YYYY:{sectionType:"year",contentType:"digit",maxLength:4},M:{sectionType:"month",contentType:"digit",maxLength:2},MM:"month",MMM:{sectionType:"month",contentType:"letter"},MMMM:{sectionType:"month",contentType:"letter"},D:{sectionType:"day",contentType:"digit",maxLength:2},DD:"day",Do:{sectionType:"day",contentType:"digit-with-letter"},d:{sectionType:"weekDay",contentType:"digit",maxLength:2},dd:{sectionType:"weekDay",contentType:"letter"},ddd:{sectionType:"weekDay",contentType:"letter"},dddd:{sectionType:"weekDay",contentType:"letter"},A:"meridiem",a:"meridiem",H:{sectionType:"hours",contentType:"digit",maxLength:2},HH:"hours",h:{sectionType:"hours",contentType:"digit",maxLength:2},hh:"hours",m:{sectionType:"minutes",contentType:"digit",maxLength:2},mm:"minutes",s:{sectionType:"seconds",contentType:"digit",maxLength:2},ss:"seconds"},E={year:"YYYY",month:"MMMM",monthShort:"MMM",dayOfMonth:"D",dayOfMonthFull:"Do",weekday:"dddd",weekdayShort:"dd",hours24h:"HH",hours12h:"hh",meridiem:"A",minutes:"mm",seconds:"ss",fullDate:"ll",keyboardDate:"L",shortDate:"MMM D",normalDate:"D MMMM",normalDateWithWeekday:"ddd, MMM D",fullTime:"LT",fullTime12h:"hh:mm A",fullTime24h:"HH:mm",keyboardDateTime:"L LT",keyboardDateTime12h:"L hh:mm A",keyboardDateTime24h:"L HH:mm"},f=["Missing UTC plugin","To be able to use UTC or timezones, you have to enable the `utc` plugin","Find more information on https://mui.com/x/react-date-pickers/timezone/#day-js-and-utc"].join(`
 `),L=["Missing timezone plugin","To be able to use timezones, you have to enable both the `utc` and the `timezone` plugin","Find more information on https://mui.com/x/react-date-pickers/timezone/#day-js-and-timezone"].join(`
 `),H=(s,r)=>r?(...o)=>s(...o).locale(r):s;class _{constructor({locale:r,formats:o}={}){this.isMUIAdapter=!0,this.isTimezoneCompatible=!0,this.lib="dayjs",this.dayjs=void 0,this.locale=void 0,this.formats=void 0,this.escapedCharacters={start:"[",end:"]"},this.formatTokenMap=U,this.setLocaleToValue=e=>{const t=this.getCurrentLocaleCode();return t===e.locale()?e:e.locale(t)},this.hasUTCPlugin=()=>typeof n.utc<"u",this.hasTimezonePlugin=()=>typeof n.tz<"u",this.isSame=(e,t,i)=>{const d=this.setTimezone(t,this.getTimezone(e));return e.format(i)===d.format(i)},this.cleanTimezone=e=>{switch(e){case"default":return;case"system":return n.tz.guess();default:return e}},this.createSystemDate=e=>{if(this.hasUTCPlugin()&&this.hasTimezonePlugin()){const t=n.tz.guess();return t!=="UTC"?n.tz(e,t):n(e)}return n(e)},this.createUTCDate=e=>{if(!this.hasUTCPlugin())throw new Error(f);return n.utc(e)},this.createTZDate=(e,t)=>{if(!this.hasUTCPlugin())throw new Error(f);if(!this.hasTimezonePlugin())throw new Error(L);const i=e!==void 0&&!e.endsWith("Z");return n(e).tz(this.cleanTimezone(t),i)},this.getLocaleFormats=()=>{const e=n.Ls,t=this.locale||"en";let i=e[t];return i===void 0&&(i=e.en),i.formats},this.adjustOffset=e=>{if(!this.hasTimezonePlugin())return e;const t=this.getTimezone(e);if(t!=="UTC"){const i=e.tz(this.cleanTimezone(t),!0);if(i.$offset===(e.$offset??0))return e;e.$offset=i.$offset}return e},this.date=(e,t="default")=>{if(e===null)return null;let i;return t==="UTC"?i=this.createUTCDate(e):t==="system"||t==="default"&&!this.hasTimezonePlugin()?i=this.createSystemDate(e):i=this.createTZDate(e,t),this.locale===void 0?i:i.locale(this.locale)},this.getInvalidDate=()=>n(new Date("Invalid date")),this.getTimezone=e=>{var t;if(this.hasTimezonePlugin()){const i=(t=e.$x)==null?void 0:t.$timezone;if(i)return i}return this.hasUTCPlugin()&&e.isUTC()?"UTC":"system"},this.setTimezone=(e,t)=>{if(this.getTimezone(e)===t)return e;if(t==="UTC"){if(!this.hasUTCPlugin())throw new Error(f);return e.utc()}if(t==="system")return e.local();if(!this.hasTimezonePlugin()){if(t==="default")return e;throw new Error(L)}return n.tz(e,this.cleanTimezone(t))},this.toJsDate=e=>e.toDate(),this.parse=(e,t)=>e===""?null:this.dayjs(e,t,this.locale,!0),this.getCurrentLocaleCode=()=>this.locale||"en",this.is12HourCycleInCurrentLocale=()=>/A|a/.test(this.getLocaleFormats().LT||""),this.expandFormat=e=>{const t=this.getLocaleFormats(),i=d=>d.replace(/(\[[^\]]+])|(MMMM|MM|DD|dddd)/g,(h,a,l)=>a||l.slice(1));return e.replace(/(\[[^\]]+])|(LTS?|l{1,4}|L{1,4})/g,(d,h,a)=>{const l=a&&a.toUpperCase();return h||t[a]||i(t[l])})},this.isValid=e=>e==null?!1:e.isValid(),this.format=(e,t)=>this.formatByString(e,this.formats[t]),this.formatByString=(e,t)=>this.dayjs(e).format(t),this.formatNumber=e=>e,this.isEqual=(e,t)=>e===null&&t===null?!0:e===null||t===null?!1:e.toDate().getTime()===t.toDate().getTime(),this.isSameYear=(e,t)=>this.isSame(e,t,"YYYY"),this.isSameMonth=(e,t)=>this.isSame(e,t,"YYYY-MM"),this.isSameDay=(e,t)=>this.isSame(e,t,"YYYY-MM-DD"),this.isSameHour=(e,t)=>e.isSame(t,"hour"),this.isAfter=(e,t)=>e>t,this.isAfterYear=(e,t)=>this.hasUTCPlugin()?!this.isSameYear(e,t)&&e.utc()>t.utc():e.isAfter(t,"year"),this.isAfterDay=(e,t)=>this.hasUTCPlugin()?!this.isSameDay(e,t)&&e.utc()>t.utc():e.isAfter(t,"day"),this.isBefore=(e,t)=>e<t,this.isBeforeYear=(e,t)=>this.hasUTCPlugin()?!this.isSameYear(e,t)&&e.utc()<t.utc():e.isBefore(t,"year"),this.isBeforeDay=(e,t)=>this.hasUTCPlugin()?!this.isSameDay(e,t)&&e.utc()<t.utc():e.isBefore(t,"day"),this.isWithinRange=(e,[t,i])=>e>=t&&e<=i,this.startOfYear=e=>this.adjustOffset(e.startOf("year")),this.startOfMonth=e=>this.adjustOffset(e.startOf("month")),this.startOfWeek=e=>this.adjustOffset(this.setLocaleToValue(e).startOf("week")),this.startOfDay=e=>this.adjustOffset(e.startOf("day")),this.endOfYear=e=>this.adjustOffset(e.endOf("year")),this.endOfMonth=e=>this.adjustOffset(e.endOf("month")),this.endOfWeek=e=>this.adjustOffset(this.setLocaleToValue(e).endOf("week")),this.endOfDay=e=>this.adjustOffset(e.endOf("day")),this.addYears=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"year"):e.add(t,"year")),this.addMonths=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"month"):e.add(t,"month")),this.addWeeks=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"week"):e.add(t,"week")),this.addDays=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"day"):e.add(t,"day")),this.addHours=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"hour"):e.add(t,"hour")),this.addMinutes=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"minute"):e.add(t,"minute")),this.addSeconds=(e,t)=>this.adjustOffset(t<0?e.subtract(Math.abs(t),"second"):e.add(t,"second")),this.getYear=e=>e.year(),this.getMonth=e=>e.month(),this.getDate=e=>e.date(),this.getHours=e=>e.hour(),this.getMinutes=e=>e.minute(),this.getSeconds=e=>e.second(),this.getMilliseconds=e=>e.millisecond(),this.setYear=(e,t)=>this.adjustOffset(e.set("year",t)),this.setMonth=(e,t)=>this.adjustOffset(e.set("month",t)),this.setDate=(e,t)=>this.adjustOffset(e.set("date",t)),this.setHours=(e,t)=>this.adjustOffset(e.set("hour",t)),this.setMinutes=(e,t)=>this.adjustOffset(e.set("minute",t)),this.setSeconds=(e,t)=>this.adjustOffset(e.set("second",t)),this.setMilliseconds=(e,t)=>this.adjustOffset(e.set("millisecond",t)),this.getDaysInMonth=e=>e.daysInMonth(),this.getWeekArray=e=>{const t=this.startOfWeek(this.startOfMonth(e)),i=this.endOfWeek(this.endOfMonth(e));let d=0,h=t;const a=[];for(;h<i;){const l=Math.floor(d/7);a[l]=a[l]||[],a[l].push(h),h=this.addDays(h,1),d+=1}return a},this.getWeekNumber=e=>e.week(),this.getYearRange=([e,t])=>{const i=this.startOfYear(e),d=this.endOfYear(t),h=[];let a=i;for(;this.isBefore(a,d);)h.push(a),a=this.addYears(a,1);return h},this.dayjs=H(n,r),this.locale=r,this.formats=m({},E,o),n.extend(z)}getDayOfWeek(r){return r.day()+1}}const W=s=>({components:{MuiLocalizationProvider:{defaultProps:{localeText:m({},s)}}}}),k={hours:"Horas",minutes:"Minutos",seconds:"Segundos",meridiem:"Meridiano"},I={previousMonth:"Mes anterior",nextMonth:"Mes siguiente",openPreviousView:"Abrir la última vista",openNextView:"Abrir la siguiente vista",calendarViewSwitchingButtonAriaLabel:s=>s==="year"?"la vista anual está abierta, cambie a la vista de calendario":"la vista de calendario está abierta, cambie a la vista anual",start:"Empezar",end:"Terminar",startDate:"Fecha inicio",startTime:"Hora inicio",endDate:"Fecha final",endTime:"Hora final",cancelButtonLabel:"Cancelar",clearButtonLabel:"Limpiar",okButtonLabel:"OK",todayButtonLabel:"Hoy",datePickerToolbarTitle:"Seleccionar fecha",dateTimePickerToolbarTitle:"Seleccionar fecha y hora",timePickerToolbarTitle:"Seleccionar hora",dateRangePickerToolbarTitle:"Seleccionar rango de fecha",clockLabelText:(s,r,o,e)=>`Seleccione ${k[s]}. ${!e&&(r===null||!o.isValid(r))?"No hay hora seleccionada":`La hora seleccionada es ${e??o.format(r,"fullTime")}`}`,hoursClockNumberText:s=>`${s} horas`,minutesClockNumberText:s=>`${s} minutos`,secondsClockNumberText:s=>`${s} segundos`,selectViewText:s=>`Seleccionar ${k[s]}`,calendarWeekNumberHeaderLabel:"Número de semana",calendarWeekNumberHeaderText:"#",calendarWeekNumberAriaLabelText:s=>`Semana ${s}`,calendarWeekNumberText:s=>`${s}`,openDatePickerDialogue:(s,r,o)=>o||s!==null&&r.isValid(s)?`Elige fecha, la fecha elegida es ${o??r.format(s,"fullDate")}`:"Elige fecha",openTimePickerDialogue:(s,r,o)=>o||s!==null&&r.isValid(s)?`Elige hora, la hora elegida es ${o??r.format(s,"fullTime")}`:"Elige hora",fieldClearLabel:"Limpiar valor",timeTableLabel:"elige hora",dateTableLabel:"elige fecha",fieldYearPlaceholder:s=>"A".repeat(s.digitAmount),fieldMonthPlaceholder:s=>s.contentType==="letter"?"MMMM":"MM",fieldDayPlaceholder:()=>"DD",fieldWeekDayPlaceholder:s=>s.contentType==="letter"?"EEEE":"EE",fieldHoursPlaceholder:()=>"hh",fieldMinutesPlaceholder:()=>"mm",fieldSecondsPlaceholder:()=>"ss",fieldMeridiemPlaceholder:()=>"aa",year:"Año",month:"Mes",day:"Dia",weekDay:"Dia de la semana",hours:"Horas",minutes:"Minutos",seconds:"Segundos",meridiem:"Meridiano",empty:"Vacío"},G=W(I);export{_ as A,B as L,G as e};
